==> Analysis/02.modelingmC.Rmd <==
---
title: "Prediction of mortality in ICU at 72 hours"
author: "Javier Quintero Sosa, Daniel Gallardo Gómez"
date: "2024-06-03"
output: html_document
---
# Model C (-24 to 48h)

# OBJECTIVE VARIABLE (TARGET)

icu_mortality_72hrs: 0 ALIVE, 1 EXPIRED

# PACKAGE LOADING (install if necessary)
                                                                                                                
```{r message=FALSE, warning=FALSE}

library(caret)
library(bigrquery)
library(bigQueryR)
library(dplyr)
library(forecast)
library(forcats)
library(ggplot2)
library(here)
library(kernlab) #required for SVM
library(knitr)
library(lubridate)
library(plotly)
library(randomForest)
library(rattle)
library(readr)
library(ROCR)
library(rpart)
library(skimr)
library(summarytools)
library(tableone)
library(table1)
library(tidyverse)
library(tidyr)
# library(xlsx)
library(xtable)

#-------------------------------------------------------------------------------

library(gbm)
library(corrplot)
library(cowplot)
library(DBI)
require(boot)
library(data.table)
library(geosphere)
library(GGally)
library(gganimate)
library(ggcorrplot)
library(gifski)
library(gridExtra)
library(hmeasure)
library(Hmisc)
library(janitor)
library(lattice)
library(magrittr)
library(MLeval)
library(mlbench)
library(MLmetrics)
library(naniar)
library(oddsratio)
library(openxlsx)
library(plotrix)
library(plyr)
library(pROC)
require(pROC)
library(psych)
library(RColorBrewer)
library(rpart.plot)
library(scales)
library(SimDesign)
library(simputation)
library(stats)
library(Tmisc)
library(visdat)
library(WVPlots)
library(xgboost)
library(modelr)
library(tidybayes)
library(ggpubr)
library(ggsci)
library(yardstick)

```


#                     EXTRACT DATAFRAME

In this section, the data, or dataframe, already analyzed and cleaned in the previous stage, is loaded.

```{r message=FALSE, warning=FALSE, paged.print=FALSE}

# 0.LOAD THE DATA

# file.choose()     # Shows the location of the file.

final_df_selected <- read_csv("C:\\Users\\jquintero24E\\Documents\\GitHub\\icu_mortality_model\\Analysis\\dataframe_mC.csv")

# Delete first column (count)

final_df_selected <- final_df_selected %>%
  select(-...1)

# 1.STUDY THE DATA

# Column names

colnames(final_df_selected)

# Check that there is no NA in the target variable.

str_glue("The number of NA in the target variable (icu_mortality_72hrs) is {sum(is.na(final_df_selected$icu_mortality_72hrs))}")

# Check the NAs in the entire dataframe.

NA_by_column <- colSums(is.na(final_df_selected))

# View(NA_by_column)

```

#             CORRECTION OF THE 0, 1 FACTOR VARIABLES

```{r}

final_df_selected <- final_df_selected %>%
  mutate(
    sex = ifelse(sex == 1, 0, 1),
    norepinephrine = ifelse(norepinephrine == 1, 0, 1),
    phenylephrine = ifelse(phenylephrine == 1, 0, 1),
    heparin = ifelse(heparin == 1, 0, 1),
    warfarin = ifelse(warfarin == 1, 0, 1),
    vasopressors_other_bin = ifelse(vasopressors_other_bin == 1, 0, 1),
    hist_diabetes_bin = ifelse(hist_diabetes_bin == 1, 0, 1),
    hist_chf_bin = ifelse(hist_chf_bin == 1, 0, 1),
    hist_copd_bin = ifelse(hist_copd_bin == 1, 0, 1),
    hist_cancer_bin = ifelse(hist_cancer_bin == 1, 0, 1),
    hist_renal_bin = ifelse(hist_renal_bin == 1, 0, 1),
    hist_other_bin = ifelse(hist_other_bin == 1, 0, 1),
    CABG = ifelse(CABG == 1, 0, 1),
    Cardiac_arrest = ifelse(Cardiac_arrest == 1, 0, 1),
    CHF = ifelse(CHF == 1, 0, 1),
    CVA = ifelse(CVA == 1, 0, 1),
    Diabetic_ketoacidosis = ifelse(Diabetic_ketoacidosis == 1, 0, 1),
    MI = ifelse(MI == 1, 0, 1),
    Rhythm_disturbance = ifelse(Rhythm_disturbance == 1, 0, 1),
    Sepsis_renal = ifelse(Sepsis_renal == 1, 0, 1),
    Sepsis_pulmonary = ifelse(Sepsis_pulmonary == 1, 0, 1),
    apache_other = ifelse(apache_other == 1, 0, 1)
  )


```



#               CHANGE THE COLUMNS BACK TO FACTOR TYPE AGAIN

```{r message=FALSE, warning=FALSE, paged.print=FALSE}

final_df_selected$icu_mortality_72hrs <- as.factor(final_df_selected$icu_mortality_72hrs)

#---
# Rename the levels of the target variable.    icu_mortality_72hrs: 0 ALIVE, 1 EXPIRED

levels(final_df_selected$icu_mortality_72hrs) <- c("Alive", "Expired")

#---

final_df_selected$hist_diabetes_bin <- as.factor(final_df_selected$hist_diabetes_bin)
final_df_selected$hist_chf_bin <- as.factor(final_df_selected$hist_chf_bin)
final_df_selected$hist_copd_bin <- as.factor(final_df_selected$hist_copd_bin)
final_df_selected$hist_cancer_bin <- as.factor(final_df_selected$hist_cancer_bin)
final_df_selected$hist_renal_bin <- as.factor(final_df_selected$hist_renal_bin)
final_df_selected$hist_other_bin <- as.factor(final_df_selected$hist_other_bin)

final_df_selected$CABG <- as.factor(final_df_selected$CABG)
final_df_selected$Cardiac_arrest <- as.factor(final_df_selected$Cardiac_arrest)
final_df_selected$CHF <- as.factor(final_df_selected$CHF)
final_df_selected$CVA <- as.factor(final_df_selected$CVA)
final_df_selected$Diabetic_ketoacidosis <- as.factor(final_df_selected$Diabetic_ketoacidosis)
final_df_selected$MI <- as.factor(final_df_selected$MI)
final_df_selected$Rhythm_disturbance <- as.factor(final_df_selected$Rhythm_disturbance)
final_df_selected$Sepsis_renal <- as.factor(final_df_selected$Sepsis_renal)
final_df_selected$Sepsis_pulmonary <- as.factor(final_df_selected$Sepsis_pulmonary)
final_df_selected$apache_other <- as.factor(final_df_selected$apache_other)

final_df_selected$sex <- as.factor(final_df_selected$sex)

final_df_selected$heparin <- as.factor(final_df_selected$heparin)
final_df_selected$norepinephrine <- as.factor(final_df_selected$norepinephrine)
final_df_selected$warfarin <- as.factor(final_df_selected$warfarin)
final_df_selected$phenylephrine <- as.factor(final_df_selected$phenylephrine)
final_df_selected$vasopressors_other_bin <- as.factor(final_df_selected$vasopressors_other_bin)

final_df_selected$region <- as.factor(final_df_selected$region)
final_df_selected$numbedscategory <- as.factor(final_df_selected$numbedscategory)
final_df_selected$teachingstatus <- as.factor(final_df_selected$teachingstatus)

final_df_selected$ethnicity <- as.factor(final_df_selected$ethnicity)

```


# REDUCTION OF VARIABLES AND SELECTION OF PATIENTS BASED ON LENGHT OF STAY IN UCI

```{r}

# Manual variable selection: I have decided to add the variables one by one to the dataframe and I have eliminated the variables that lowered the performance of the model (decreased the F1).


# Excluded Variables
# hist_other_bin F1 drops to 0,583
# spo2_clear F1 drops a little
# region has missing values
# numbedscategory has missing values
# apache_other could add noise.


# Essential Variables (if they are eliminated, the F1 decreases)
# prev_dept_los_hrs
# potassium_clear
# hemoglobin_clear
# final_charlson_score
# Cardiac_arrest 

## I have made a dataframe only with the "Essential" variables, but it is a failure: F1: 0.4066 Successes: 32.749%

## Added: region, numbedscategory, teachingstatus (it reduces success)

final_df_selected <- final_df_selected %>%
  select(patientunitstayid, LOS, icu_mortality_72hrs, age, sex, ethnicity, prev_dept_los_hrs, bicarbonate_clear, sodium_clear, respiratoryrate_clear, BUN_clear, nibp_diastolic_clear, nibp_systolic_clear, nibp_mean_clear, temperature_clear, glucose_clear, potassium_clear, hemoglobin_clear, calcium_clear, norepinephrine, phenylephrine, heparin, warfarin, heartrate_clear, vasopressors_other_bin, hist_diabetes_bin, hist_chf_bin, hist_copd_bin, hist_cancer_bin, hist_renal_bin, final_charlson_score, CABG, Cardiac_arrest, CHF, CVA, Diabetic_ketoacidosis, MI, Rhythm_disturbance, Sepsis_renal, Sepsis_pulmonary, teachingstatus)


# Delete "LOS" column (the same for Models A, B and C).

final_df_selected <- final_df_selected %>%
  select(-LOS)

# stview(dfSummary(final_df_selected))

```

Factor variables:

icu_mortality_72hrs (Alive: 0 / Expired: 1)
Sex (Male: 0 / Female: 1)
ethnicity (African American: 0 / Asian: 1 / Caucasian: 2 / Hispanic: 3 / Native American: 4 / Other/Unknown: 5)
teachingstatus (FALSE: 0   TRUE: 1)



#                   DIVISION OF THE DF IN TRAIN AND TEST

To train various types of models that can predict a patient's survival based on the data, said data will be divided into two sets: training (train) and testing (test).

```{r message=FALSE, warning=FALSE}


# 0. CHECK BALANCE OR IMBALANCE OF THE DATAFRAME

round(prop.table(table(final_df_selected$icu_mortality_72hrs)),2)

# 1. TRAIN-TEST SELECTION

# Set the seed to make our partition reproducible.
set.seed(13579)

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_selected$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train = final_df_selected[idx_train, ]
data_test  = final_df_selected[-idx_train, ]

# 2. Create a summarytools (train and test)

# stview(dfSummary(data_train))
# stview(dfSummary(data_test))

# 3. SAVE DIVISION TO CSV

   write.csv(data_test, "data_test_C.csv")
   write.csv(data_train, "data_train.csv")

```

# ESTABLISH THE FORMULA FOR THE MODEL

A formula of type icu_mortality_72hrs ~ predictor1 + predictor2 + ... is used to train the model.

Only the variables of interest must be selected, as well as the patient ID removed (except in data_test).


```{r message=FALSE, warning=FALSE}

# 1º REMOVE ID FROM data_train AND FROM ORIGINAL DATAFRAME (data_test must have ID)

final_df_selected_without_ID <- final_df_selected %>%
  select(-patientunitstayid)

data_train_without_ID <- data_train %>%
  select(-patientunitstayid)

# At this point you can remove unwanted columns for modeling.

# 2º. MAKE THE MODEL (without including the ID as a variable)

outcome_model <- "icu_mortality_72hrs"                                   
exposures_model <- names(data_train_without_ID)                   
exposures_model <- exposures_model[-which(exposures_model == outcome_model)]   

# Model formula

outcome_and_exposure <- as.formula(paste(outcome_model, "~", paste(exposures_model, collapse = " + "))) 

```

# TABLE FOR COMPARISON TRAIN - TEST

It is verified that the variables included in the model are represented approximately equally in the training and test sets.

```{r message=FALSE, warning=FALSE}

tmp = as.data.frame(final_df_selected_without_ID)
tmp$trainTest = "TestSet"
tmp$trainTest[idx_train] = "TrainingSet"
table1(as.formula(paste0(" ~ ",paste(exposures_model, collapse = " + ")," | trainTest")), data=tmp, overall=F)

```


# HYPERPARAMETER TUNING

This adjustment randomly tests all study models and chooses the one with the best results.

The first thing to do is set the fitcontrol (establish how to evaluate each model).

```{r message=FALSE, warning=FALSE}

fitControl <- trainControl(method = "repeatedcv",
                           number = 10,
                           repeats = 1,
                           classProbs = TRUE,
                           summaryFunction = multiClassSummary,
                           verboseIter = F,
                           search = "random",
                           sampling = "down")

```

# MODEL TRAINING AND BEST MODEL SELECTION

The different models will be trained:

gbm: Gradient Boosting Machine
rf: Random Forest
xgbTree: eXtreme Gradient Boosting
LogitBoost: Boosted Logistic Regression
lda: Linear Discriminant Analysis
kernelpls: Kernel Partial Least Squares
ranger: Random forests
Linda
Rborist: Random Forest

The train() function is used to train the models.

```{r message=FALSE, warning=FALSE}

# Set the seed to make our partition reproducible.
set.seed(123)

# Train the different models.

h_start_gbm <- Sys.time()                          

fit_gbm <- train(outcome_and_exposure,                                  # 16 min
                 data = data_train_without_ID %>% as.data.frame(),
                 method = "gbm",               
                 trControl = fitControl,        
                 verbose = F,                   
                 na.action = na.omit,
                 metric = "F1")                 

h_start_rf <- Sys.time()                           

fit_rf <- train(outcome_and_exposure,                                   # 6 min
                data = data_train_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

h_start_ada <- Sys.time()                           

fit_ada <- train(outcome_and_exposure,                                  # 11 min  
                data = data_train_without_ID %>% as.data.frame(),
                method = "ada",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

h_start_xgb <- Sys.time()                          

fit_xgb <- train(outcome_and_exposure,                                  # 2 min  
                 data = data_train_without_ID %>% as.data.frame(),
                 method = "xgbTree",
                 trControl = fitControl,
                 verbose = F,
                 na.action = na.omit,
                 metric = "F1")

h_start_lr <- Sys.time()                          

fit_lr <- train(outcome_and_exposure,                                   # 15 s  
                data = data_train_without_ID %>% as.data.frame(),
                method = "LogitBoost",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

h_start_lda <- Sys.time()                                         

fit_lda <- train(outcome_and_exposure,                                  # 8 s 
                data = data_train_without_ID %>% as.data.frame(),          
                method = "lda",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

h_start_kernelpls <- Sys.time()                  

fit_kernelpls <- train(outcome_and_exposure,                            # 19 s
                data = data_train_without_ID %>% as.data.frame(),
                method = "kernelpls",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

h_start_ranger <- Sys.time() 

fit_ranger <- train(outcome_and_exposure,                               # 1 min  
                data = data_train_without_ID %>% as.data.frame(),
                method = "ranger",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

h_start_Linda <- Sys.time() 

fit_Linda <- train(outcome_and_exposure,                                # 7 s
                data = data_train_without_ID %>% as.data.frame(),      
                method = "Linda",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

h_start_cforest <- Sys.time() 

# Better to run this model separately (sometimes it crashes)

fit_cforest <- train(outcome_and_exposure,                              # 14 min  
                data = data_train_without_ID %>% as.data.frame(),    
                method = "Rborist",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

h_end <- Sys.time()

# Total modeling time

# Minutes
minutes_modeling <- as.numeric(difftime(h_end, h_start_gbm, units = "mins"))

# Hours
h_modeling <- as.numeric(difftime(h_end, h_start_gbm, units = "hours"))


#---------------------------------------------------------------------------------------------------------------
#---------------------------------------------------------------------------------------------------------------

# SAVE THE MODELS

saveRDS(fit_gbm, file = "gbm_model.Rda")
saveRDS(fit_rf, file = "rf_model.Rda")
saveRDS(fit_ada, file = "ada_model.Rda")
saveRDS(fit_xgb, file = "xgb_model.Rda")
saveRDS(fit_lr, file = "lr_model.Rda")
saveRDS(fit_lda, file = "lda_model.Rda")
saveRDS(fit_kernelpls, file = "kernelpls_model.Rda")
saveRDS(fit_ranger, file = "ranger_model.Rda")
saveRDS(fit_Linda, file = "Linda_model.Rda")
saveRDS(fit_cforest, file = "cforest_model.Rda")

# UPLOAD THE MODELS

# fit_gbm <- readRDS("gbm_model.Rda")
# fit_rf <- readRDS("rf_model.Rda")
# fit_ada <- readRDS("ada_model.Rda")
# fit_xgb <- readRDS("xgb_model.Rda")
# fit_lr <- readRDS("lr_model.Rda")
# fit_lda <- readRDS("lda_model.Rda")
# fit_kernelpls <- readRDS("kernelpls_model.Rda")
# fit_ranger <- readRDS("ranger_model.Rda")
# fit_Linda <- readRDS("Linda_model.Rda")
# fit_cforest <- readRDS("cforest_model.Rda")

#---------------------------------------------------------------------------------------------------------------
#---------------------------------------------------------------------------------------------------------------

# STATISTICS SUMMARY

resamps <- resamples(list(fit_gbm = fit_gbm,
                          fit_rf  = fit_rf,
                          fit_ada  = fit_ada,
                          fit_xgb = fit_xgb,
                          fit_lr  = fit_lr,
                          fit_lda = fit_lda,
                          fit_kernelpls = fit_kernelpls,
                          fit_ranger = fit_ranger,
                          fit_Linda = fit_Linda,
                          fit_cforest= fit_cforest
                          ))                 
summary_resamps = as.data.frame(summary(resamps)$statistics)   

summary_resamps %>% t() %>% kable()   

#---------------------------------------------------------------------------------------------------------------
#---------------------------------------------------------------------------------------------------------------

# SELECT THE HIGHEST PERFORMANCE MODEL

# Save the model with the highest performance. We will focus on "Balanced_Accuracy.Mean"

best_performing_model<-get(
  rownames(summary_resamps)[which(summary_resamps$Balanced_Accuracy.Mean == max(summary_resamps$Balanced_Accuracy.Mean))])

# Extract model name

best_performing_model_name<-best_performing_model$method  

# Print the name.

str_glue("The best model, only based on Balanced Accuracy, is {best_performing_model_name}")

# Select the model

fit_best <- fit_xgb

#---------------------------------------------------------------------------------------------------------------
#---------------------------------------------------------------------------------------------------------------

# PREDICTOR EVALUATION. (From now on data_test is used)

# Predictions are calculated for all test set data in all models. It is defined Alive=0 and Expired=1

# CREATE THE getFinalPredictions FUNCTION.

getFinalPredictions = function(fit_best, data_test, prob_threshold=0.4){               
                                                                                           
  prediction_probabilities_tmp = predict(fit_best, newdata=data_test, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test$patientunitstayid,                
                                 trueStatus = data_test$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))  # If "Alive" (alive probability) exceeds the threshold, enter 0
  
  final_predictions_tmp
}

# End of function

#-------------------------------------------------------------------------------

# Use the getFinalPredictions function on the best model and all the others.

final_predictions = getFinalPredictions(fit_best, data_test)
final_predictions_gbm = getFinalPredictions(fit_gbm, data_test)
final_predictions_rf = getFinalPredictions(fit_rf, data_test)
final_predictions_ada = getFinalPredictions(fit_ada, data_test)
final_predictions_xgb = getFinalPredictions(fit_xgb, data_test)
final_predictions_lr = getFinalPredictions(fit_lr, data_test)
final_predictions_lda = getFinalPredictions(fit_lda, data_test)
final_predictions_kernelpls = getFinalPredictions(fit_kernelpls, data_test)
final_predictions_ranger = getFinalPredictions(fit_ranger, data_test)
final_predictions_Linda = getFinalPredictions(fit_Linda, data_test)
final_predictions_cforest = getFinalPredictions(fit_cforest, data_test)

# We evaluate the Balanced Accuracy, the F1 parameter and the confusion matrix, and select the model that meets the most criteria.

```


# CONFUSION MATRIX OF ALL MODELS

```{r}

# gbm CONFUSION MATRIX

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_gbm$obs <- as.factor(final_predictions_gbm$obs)

# mortality_prediction_B
final_predictions_gbm$pred <- as.factor(final_predictions_gbm$pred)

# confusion matrix

matrix_gbm <- confusionMatrix(final_predictions_gbm$pred, final_predictions_gbm$obs, mode = "everything")

#---

# rf CONFUSION MATRIX

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_rf$obs <- as.factor(final_predictions_rf$obs)

# mortality_prediction_B
final_predictions_rf$pred <- as.factor(final_predictions_rf$pred)

# confusion matrix

matrix_rf <- confusionMatrix(final_predictions_rf$pred, final_predictions_rf$obs, mode = "everything")

#---

# ada CONFUSION MATRIX

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_ada$obs <- as.factor(final_predictions_ada$obs)

# mortality_prediction_B
final_predictions_ada$pred <- as.factor(final_predictions_ada$pred)

# confusion matrix

matrix_ada <- confusionMatrix(final_predictions_ada$pred, final_predictions_ada$obs, mode = "everything")

#---

# xgb CONFUSION MATRIX

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_xgb$obs <- as.factor(final_predictions_xgb$obs)

# mortality_prediction_B
final_predictions_xgb$pred <- as.factor(final_predictions_xgb$pred)

# confusion matrix

matrix_xgb <- confusionMatrix(final_predictions_xgb$pred, final_predictions_xgb$obs, mode = "everything")

#---

# lr CONFUSION MATRIX

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_lr$obs <- as.factor(final_predictions_lr$obs)

# mortality_prediction_B
final_predictions_lr$pred <- as.factor(final_predictions_lr$pred)

# confusion matrix

matrix_lr <- confusionMatrix(final_predictions_lr$pred, final_predictions_lr$obs, mode = "everything")

#---

# lda CONFUSION MATRIX

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_lda$obs <- as.factor(final_predictions_lda$obs)

# mortality_prediction_B
final_predictions_lda$pred <- as.factor(final_predictions_lda$pred)

# confusion matrix

matrix_lda <- confusionMatrix(final_predictions_lda$pred, final_predictions_lda$obs, mode = "everything")

#---

# kernelpls CONFUSION MATRIX

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_kernelpls$obs <- as.factor(final_predictions_kernelpls$obs)

# mortality_prediction_B
final_predictions_kernelpls$pred <- as.factor(final_predictions_kernelpls$pred)

# confusion matrix

matrix_kernelpls <- confusionMatrix(final_predictions_kernelpls$pred, final_predictions_kernelpls$obs, mode = "everything")

#---

# ranger CONFUSION MATRIX

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_ranger$obs <- as.factor(final_predictions_ranger$obs)

# mortality_prediction_B
final_predictions_ranger$pred <- as.factor(final_predictions_ranger$pred)

# confusion matrix

matrix_ranger <- confusionMatrix(final_predictions_ranger$pred, final_predictions_ranger$obs, mode = "everything")

#---

# Linda CONFUSION MATRIX

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_Linda$obs <- as.factor(final_predictions_Linda$obs)

# mortality_prediction_B
final_predictions_Linda$pred <- as.factor(final_predictions_Linda$pred)

# confusion matrix

matrix_Linda <- confusionMatrix(final_predictions_Linda$pred, final_predictions_Linda$obs, mode = "everything")

#---

# cforest CONFUSION MATRIX

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_cforest$obs <- as.factor(final_predictions_cforest$obs)

# mortality_prediction_B
final_predictions_cforest$pred <- as.factor(final_predictions_cforest$pred)

# confusion matrix

matrix_cforest <- confusionMatrix(final_predictions_cforest$pred, final_predictions_cforest$obs, mode = "everything")

```


# COMPARATIVE TABLE

```{r}

# PRECISION

gbm_Precision <- matrix_gbm$table[1, 1] / sum(matrix_gbm$table[1, ])
rf_Precision <- matrix_rf$table[1, 1] / sum(matrix_rf$table[1, ])
ada_Precision <- matrix_ada$table[1, 1] / sum(matrix_ada$table[1, ])
xgb_Precision <- matrix_xgb$table[1, 1] / sum(matrix_xgb$table[1, ])
lr_Precision <- matrix_lr$table[1, 1] / sum(matrix_lr$table[1, ])
lda_Precision <- matrix_lda$table[1, 1] / sum(matrix_lda$table[1, ])
kernelpls_Precision <- matrix_kernelpls$table[1, 1] / sum(matrix_kernelpls$table[1, ])
ranger_Precision <- matrix_ranger$table[1, 1] / sum(matrix_ranger$table[1, ])
Linda_Precision <- matrix_Linda$table[1, 1] / sum(matrix_Linda$table[1, ])
cforest_Precision <- matrix_cforest$table[1, 1] / sum(matrix_cforest$table[1, ])


# RECALL

gbm_recall <- matrix_gbm$table[1, 1] / sum(matrix_gbm$table[, 1])
rf_recall <- matrix_rf$table[1, 1] / sum(matrix_rf$table[, 1])
ada_recall <- matrix_ada$table[1, 1] / sum(matrix_ada$table[, 1])
xgb_recall <- matrix_xgb$table[1, 1] / sum(matrix_xgb$table[, 1])
lr_recall <- matrix_lr$table[1, 1] / sum(matrix_lr$table[, 1])
lda_recall <- matrix_lda$table[1, 1] / sum(matrix_lda$table[, 1])
kernelpls_recall <- matrix_kernelpls$table[1, 1] / sum(matrix_kernelpls$table[, 1])
ranger_recall <- matrix_ranger$table[1, 1] / sum(matrix_ranger$table[, 1])
Linda_recall <- matrix_Linda$table[1, 1] / sum(matrix_Linda$table[, 1])
cforest_recall <- matrix_cforest$table[1, 1] / sum(matrix_cforest$table[, 1])


# PARAMETER F1.

gbm_F1 <-F1_Score(final_predictions_gbm$obs, final_predictions_gbm$pred)
rf_F1 <-F1_Score(final_predictions_rf$obs, final_predictions_rf$pred)
ada_F1 <-F1_Score(final_predictions_ada$obs, final_predictions_ada$pred)
xgb_F1 <-F1_Score(final_predictions_xgb$obs, final_predictions_xgb$pred)
lr_F1 <-F1_Score(final_predictions_lr$obs, final_predictions_lr$pred)
lda_F1 <-F1_Score(final_predictions_lda$obs, final_predictions_lda$pred)
kernelpls_F1 <-F1_Score(final_predictions_kernelpls$obs, final_predictions_kernelpls$pred)
ranger_F1 <-F1_Score(final_predictions_ranger$obs, final_predictions_ranger$pred)
Linda_F1 <-F1_Score(final_predictions_Linda$obs, final_predictions_Linda$pred)
cforest_F1 <-F1_Score(final_predictions_cforest$obs, final_predictions_cforest$pred)


# BALANCED ACCURACY

gbm_BA <- matrix_gbm$byClass['Balanced Accuracy']
rf_BA <- matrix_rf$byClass['Balanced Accuracy']
ada_BA <- matrix_ada$byClass['Balanced Accuracy']
xgb_BA <- matrix_xgb$byClass['Balanced Accuracy']
lr_BA <- matrix_lr$byClass['Balanced Accuracy']
lda_BA <- matrix_lda$byClass['Balanced Accuracy']
kernelpls_BA <- matrix_kernelpls$byClass['Balanced Accuracy']
ranger_BA <- matrix_ranger$byClass['Balanced Accuracy']
Linda_BA <- matrix_Linda$byClass['Balanced Accuracy']
cforest_BA <- matrix_cforest$byClass['Balanced Accuracy']

#-------------------------------------------------------------------------------

# CREATE THE TABLE

# Models
Models <- c("lr", "cforest", "lda", "kernelpls", "Linda", "gbm", "xgb", "ranger", "rf", "ada")

# Precision
Precision <- round(c(lr_Precision, cforest_Precision, lda_Precision, kernelpls_Precision, Linda_Precision, gbm_Precision, xgb_Precision, ranger_Precision, rf_Precision, ada_Precision),3)

# Recall
Recall <- round(c(lr_recall, cforest_recall, lda_recall, kernelpls_recall, Linda_recall, gbm_recall, xgb_recall, ranger_recall, rf_recall, ada_recall),3)

# F1 score
F1 <- round(c(lr_F1, cforest_F1, lda_F1, kernelpls_F1, Linda_F1, gbm_F1, xgb_F1, ranger_F1, rf_F1, ada_F1),3)

# Balanced Accuracy

Balanced_Accuracy <- round(c(lr_BA, cforest_BA, lda_BA, kernelpls_BA, Linda_BA, gbm_BA, xgb_BA, ranger_BA, rf_BA, ada_BA),3)


# Dataframe
comparative_table <- data.frame(Models = Models, Balanced_Accuracy = Balanced_Accuracy, Precision = Precision, Recall = Recall, F1 = F1)

print(comparative_table)

comparative_table_ordered <- comparative_table[order(comparative_table$Balanced_Accuracy, decreasing = TRUE),]

print(comparative_table_ordered)

#-------------------------------------------------------------------------------

# MODEL UNCERTAINTY

# This could help to choose the best model.

# y: mean value; ymin: CI 95% lower bound; ymax:CI 95% upper bound

uncertainty_lr <- mean_qi(fit_lr$resample$F1)
uncertainty_cforest <- mean_qi(fit_cforest$resample$F1)
uncertainty_lda <- mean_qi(fit_lda$resample$F1)
uncertainty_kernelpls <- mean_qi(fit_kernelpls$resample$F1)
uncertainty_Linda <- mean_qi(fit_Linda$resample$F1)
uncertainty_gbm <- mean_qi(fit_gbm$resample$F1)
uncertainty_xgb <- mean_qi(fit_xgb$resample$F1)
uncertainty_ranger <- mean_qi(fit_ranger$resample$F1)
uncertainty_rf <- mean_qi(fit_rf$resample$F1)
uncertainty_ada <- mean_qi(fit_ada$resample$F1)


# Models
Models <- c("lr", "cforest", "lda", "kernelpls", "Linda", "gbm", "xgb", "ranger", "rf", "ada")

# F1 mean

mean_F1_lr <- round(pull(uncertainty_lr, y),3)
mean_F1_cforest <- round(pull(uncertainty_cforest, y),3)
mean_F1_lda <- round(pull(uncertainty_lda, y),3)
mean_F1_kernelpls <- round(pull(uncertainty_kernelpls, y),3)
mean_F1_Linda <- round(pull(uncertainty_Linda, y),3)
mean_F1_gbm <- round(pull(uncertainty_gbm, y),3)
mean_F1_xgb <- round(pull(uncertainty_xgb, y),3)
mean_F1_ranger <- round(pull(uncertainty_ranger, y),3)
mean_F1_rf <- round(pull(uncertainty_rf, y),3)
mean_F1_ada <- round(pull(uncertainty_ada, y),3)

F1_mean <- c(mean_F1_lr, mean_F1_cforest, mean_F1_lda, mean_F1_kernelpls, mean_F1_Linda, mean_F1_gbm, mean_F1_xgb, mean_F1_ranger, mean_F1_rf, mean_F1_ada)


# F1 lower limit

lower_F1_lr <- round(pull(uncertainty_lr, ymin),3)
lower_F1_cforest <- round(pull(uncertainty_cforest, ymin),3)
lower_F1_lda <- round(pull(uncertainty_lda, ymin),3)
lower_F1_kernelpls <- round(pull(uncertainty_kernelpls, ymin),3)
lower_F1_Linda <- round(pull(uncertainty_Linda, ymin),3)
lower_F1_gbm <- round(pull(uncertainty_gbm, ymin),3)
lower_F1_xgb <- round(pull(uncertainty_xgb, ymin),3)
lower_F1_ranger <- round(pull(uncertainty_ranger, ymin),3)
lower_F1_rf <- round(pull(uncertainty_rf, ymin),3)
lower_F1_ada <- round(pull(uncertainty_ada, ymin),3)

lower_limit <- c(lower_F1_lr, lower_F1_cforest, lower_F1_lda, lower_F1_kernelpls, lower_F1_Linda, lower_F1_gbm, lower_F1_xgb, lower_F1_ranger, lower_F1_rf, lower_F1_ada)

# F1 upper limit

upper_F1_lr <- round(pull(uncertainty_lr, ymax),3)
upper_F1_cforest <- round(pull(uncertainty_cforest, ymax),3)
upper_F1_lda <- round(pull(uncertainty_lda, ymax),3)
upper_F1_kernelpls <- round(pull(uncertainty_kernelpls, ymax),3)
upper_F1_Linda <- round(pull(uncertainty_Linda, ymax),3)
upper_F1_gbm <- round(pull(uncertainty_gbm, ymax),3)
upper_F1_xgb <- round(pull(uncertainty_xgb, ymax),3)
upper_F1_ranger <- round(pull(uncertainty_ranger, ymax),3)
upper_F1_rf <- round(pull(uncertainty_rf, ymax),3)
upper_F1_ada <- round(pull(uncertainty_ada, ymax),3)

upper_limit <- c(upper_F1_lr, upper_F1_cforest, upper_F1_lda, upper_F1_kernelpls, upper_F1_Linda, upper_F1_gbm, upper_F1_xgb, upper_F1_ranger, upper_F1_rf, upper_F1_ada)


# Dataframe
uncertainty_table <- data.frame(Models = Models, F1_mean = F1_mean, lower_limit = lower_limit, upper_limit = upper_limit)

print(uncertainty_table)

```
The model gbm has the best balanced accuracy, but there are other effective models.All models have a high performance F1, so to evaluate other models the confusion matrix and Specificity can be considered. Models rf and xgb are considered to be effective models.


# TRAINING ONLY THE SELECTED MODELS

The selected models are: gbm, rf and xgb.

## Model 1 (gbm) training

```{r}

# Training and predictions ONLY of the created model

set.seed(123)

fit_gbm <- train(outcome_and_exposure,                              
                 data = data_train_without_ID %>% as.data.frame(),
                 method = "gbm",               
                 trControl = fitControl,        
                 verbose = F,                   
                 na.action = na.omit,
                 metric = "F1") 

#---

# SAVE THE MODEL

# You can save the model so you don't have to run it again.

saveRDS(fit_gbm, file = "gbm_model.Rda")

# UPLOAD THE MODEL

# If you have run model training previously you can load the saved model to save time.

fit_best_1 <- readRDS("gbm_model.Rda")

#---

# CREATE THE getFinalPredictions FUNCTION.

getFinalPredictions = function(fit_best_1, data_test, prob_threshold=0.4){     
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_1, newdata=data_test, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test$patientunitstayid,                
                                 trueStatus = data_test$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))   # If "Alive" (alive probability) exceeds the threshold, enter 0
  
  final_predictions_tmp
}

# End of function

final_predictions_best_1 = getFinalPredictions(fit_best_1, data_test)

```


## Model 2 (rf) training

```{r}

# Training and predictions ONLY of the created model

set.seed(123)

fit_rf <- train(outcome_and_exposure,                               
                data = data_train_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

#---

# SAVE THE MODEL

# You can save the model so you don't have to run it again.

saveRDS(fit_rf, file = "rf_model.Rda")

# UPLOAD THE MODEL

# If you have run model training previously you can load the saved model to save time.

fit_best_2 <- readRDS("rf_model.Rda")

#---

# CREATE THE getFinalPredictions FUNCTION.

getFinalPredictions = function(fit_best_2, data_test, prob_threshold=0.4){     
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_2, newdata=data_test, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test$patientunitstayid,                
                                 trueStatus = data_test$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))   # If "Alive" (alive probability) exceeds the threshold, enter 0
  
  final_predictions_tmp
}

# End of function

final_predictions_best_2 = getFinalPredictions(fit_best_2, data_test)

```


## Model 3 (xgb) training

```{r}

# Training and predictions ONLY of the created model

set.seed(123)

fit_xgb <- train(outcome_and_exposure,                             
                 data = data_train_without_ID %>% as.data.frame(),
                 method = "xgbTree",
                 trControl = fitControl,
                 verbose = F,
                 na.action = na.omit,
                 metric = "F1")

#---

# SAVE THE MODEL

# You can save the model so you don't have to run it again.

saveRDS(fit_xgb, file = "xgb_model.Rda")

# UPLOAD THE MODEL

# If you have run model training previously you can load the saved model to save time.

fit_best_3 <- readRDS("xgb_model.Rda")

#---

# CREATE THE getFinalPredictions FUNCTION.

getFinalPredictions = function(fit_best_3, data_test, prob_threshold=0.4){     
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_3, newdata=data_test, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test$patientunitstayid,                
                                 trueStatus = data_test$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))   # If "Alive" (alive probability) exceeds the threshold, enter 0
  
  final_predictions_tmp
}

# End of function

final_predictions_best_3 = getFinalPredictions(fit_best_3, data_test)

```


# CONFUSION MATRIX OF THE ALREADY CHOSEN MODELS

```{r}

#                     CONFUSION MATRIX (selected model 1)

str_glue("-------------------------------- Model 1: gbm")

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_best_1$obs <- as.factor(final_predictions_best_1$obs)

# mortality_prediction_B
final_predictions_best_1$pred <- as.factor(final_predictions_best_1$pred)

# confusion matrix

confusionMatrix(final_predictions_best_1$pred, final_predictions_best_1$obs, mode = "everything")


#-------------------------------------------------------------------------------

#                     CONFUSION MATRIX (selected model 2)

str_glue("-------------------------------- Model 2: rf")

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_best_2$obs <- as.factor(final_predictions_best_2$obs)

# mortality_prediction_B
final_predictions_best_2$pred <- as.factor(final_predictions_best_2$pred)

# confusion matrix

confusionMatrix(final_predictions_best_2$pred, final_predictions_best_2$obs, mode = "everything")


#-------------------------------------------------------------------------------

#                     CONFUSION MATRIX (selected model 3)

str_glue("-------------------------------- Model 3: xgb")

# The dependent variable must be of type factor.

# icu_mortality_72hrs
final_predictions_best_3$obs <- as.factor(final_predictions_best_3$obs)

# mortality_prediction_B
final_predictions_best_3$pred <- as.factor(final_predictions_best_3$pred)

# confusion matrix

confusionMatrix(final_predictions_best_3$pred, final_predictions_best_3$obs, mode = "everything")

```

# COMPARING MODEL PERFORMANCE APPLYING THRESHOLD-MOVING FOR IMBALANCE CLASSIFICATION

## Selected model 1

```{r}

# For imbalance classification tasks, it is recommended trying to move the probability threshold (by default 0.5) to achieve a better model performance

# Applying a probability threshold of 0.4 (i.e., if the probability of living is 0.4 or more, the status of this patient is classified as "Alive")
final_predictions_0.5 = getFinalPredictions(fit_best_1, data_test, prob_threshold = 0.5)

# transforming outcomes as factors

final_predictions_0.5$obs <- as.factor(final_predictions_0.5$obs)

levels(final_predictions_0.5$obs)

final_predictions_0.5$obs <- as.factor(final_predictions_0.5$obs)

# mortality_prediction_B
final_predictions_0.5$pred <- factor(final_predictions_0.5$pred)

levels(final_predictions_0.5$pred)

final_predictions_best_1$pred <- factor(final_predictions_best_1$pred)

# Comparing both Confusion Matrices
str_glue("-------------------------------- Threshold: 0.40")

## matrix applying threshold of 0.40
confusionMatrix(final_predictions_best_1$pred, final_predictions_best_1$obs, mode = "everything")

str_glue("-------------------------------- Threshold: 0.50")
## matrix applying by default threshold of 0.50
confusionMatrix(final_predictions_0.5$pred, final_predictions_0.5$obs, mode = "everything")
```

The same goes for the rest of the models.


# PRECISION-RECALL CURVE

In this section the optimal threshold is calculated.

```{r}

# (Selected model 1)

pr_curve(final_predictions_best_1, obs, Alive) %>%
    ggplot(aes(x = recall, y = precision)) +
    geom_path() +
    theme_bw() +
  geom_point(data = pr_curve(final_predictions_best_1, obs, Alive) %>% filter(round(.threshold, 2) %in% c(0.3, 0.4, 0.5)), size = 3) + labs(x = "Recall", y = "Precision") +
  geom_label(aes(label = round(.threshold, 2)),
                                                                                                                                                                              #fill = factor(.threshold),
                                                                                                                                                                          col = "white", fill = "black", data = pr_curve(final_predictions_best_1, obs, Alive) %>% filter(round(.threshold, 2) %in% c(0.30, 0.40, 0.50)), vjust = 0, nudge_y = 0.0015) + theme(legend.position = "none") +
labs(title = "PRECISION-RECALL CURVE OF THE MODEL 1 (gbm)")

#-------------------------------------------------------------------------------

# (Selected model 2)

pr_curve(final_predictions_best_2, obs, Alive) %>%
    ggplot(aes(x = recall, y = precision)) +
    geom_path() +
    theme_bw() +
  geom_point(data = pr_curve(final_predictions_best_2, obs, Alive) %>% filter(round(.threshold, 2) %in% c(0.3, 0.4, 0.5)), size = 3) + labs(x = "Recall", y = "Precision") +
  geom_label(aes(label = round(.threshold, 2)),
                                                                                                                                                                              #fill = factor(.threshold),
                                                                                                                                                                          col = "white", fill = "black", data = pr_curve(final_predictions_best_2, obs, Alive) %>% filter(round(.threshold, 2) %in% c(0.30, 0.40, 0.50)), vjust = 0, nudge_y = 0.0015) + theme(legend.position = "none") +
labs(title = "PRECISION-RECALL CURVE OF THE MODEL 2 (rf)")

#-------------------------------------------------------------------------------

# (Selected model 3)

pr_curve(final_predictions_best_3, obs, Alive) %>%
    ggplot(aes(x = recall, y = precision)) +
    geom_path() +
    theme_bw() +
  geom_point(data = pr_curve(final_predictions_best_3, obs, Alive) %>% filter(round(.threshold, 2) %in% c(0.3, 0.4, 0.5)), size = 3) + labs(x = "Recall", y = "Precision") +
  geom_label(aes(label = round(.threshold, 2)),
                                                                                                                                                                              #fill = factor(.threshold),
                                                                                                                                                                          col = "white", fill = "black", data = pr_curve(final_predictions_best_3, obs, Alive) %>% filter(round(.threshold, 2) %in% c(0.30, 0.40, 0.50)), vjust = 0, nudge_y = 0.0015) + theme(legend.position = "none") +
labs(title = "PRECISION-RECALL CURVE OF THE MODEL 3 (xgb)")

```
As anticipated in the previous section, the optimal threshold is 0.40


# CHECK THE IMPORTANCE OF THE MODEL VARIABLES.

A graphic representation is made of the importance of each variable in the model.

```{r message=FALSE, warning=FALSE}

# Calculate the relative importance of the model variables.

importance1 <- varImp(fit_best_1)
importance2 <- varImp(fit_best_2)
importance3 <- varImp(fit_best_3)


# ggplot

ggplot(importance1, aes(x = Overall, y = reorder(rownames(importance), Overall))) +
  geom_bar(stat = "identity", width = 0.5, fill = "blue") +
  labs(title = "IMPORTANCE OF THE MODEL 1 (gbm) VARIABLES",
       x = "Variables",
       y = "Relative importance (%)") +
  theme_minimal() +
  theme(axis.text.y = element_text(size = 7)) +
  coord_flip() 

#---

ggplot(importance2, aes(x = Overall, y = reorder(rownames(importance), Overall))) +
  geom_bar(stat = "identity", width = 0.5, fill = "blue") +
  labs(title = "IMPORTANCE OF THE MODEL 2 (rf) VARIABLES",
       x = "Variables",
       y = "Relative importance (%)") +
  theme_minimal() +
  theme(axis.text.y = element_text(size = 7)) +
  coord_flip() 

#---

ggplot(importance3, aes(x = Overall, y = reorder(rownames(importance), Overall))) +
  geom_bar(stat = "identity", width = 0.5, fill = "blue") +
  labs(title = "IMPORTANCE OF THE MODEL 3 (xgb) VARIABLES",
       x = "Variables",
       y = "Relative importance (%)") +
  theme_minimal() +
  theme(axis.text.y = element_text(size = 7)) +
  coord_flip() 


# norepinefrina, respiratoryrate_vlear, vasopresors other, BUN

```

# SUBGROUPS PREDICTIONS (FAIRNESS)

This analysis will only be carried out for the model with the highest Balanced Accuracy.


## Option A: We use the trained model to predict by subgroups.

As will be seen below, in the section "OUTCOMES: PREDICTIONS USING `data_test`", the most precise model is rf (Model 2), so that model will be the one used in this section.

```{r}

# Creation of the final_predictions data joining our data test

final_predictions <- final_predictions_best_2


final_predictions <- final_predictions %>%
  left_join(data_test %>% dplyr::rename(ID = patientunitstayid), by = "ID") %>%
  
#-------------------------------------------------------------------------------

# CEATION OF NECESSARY GROUPS  

  # create age groups
  
  mutate(age_class = case_when(
    age > 0 & age < 10 ~ "childhood",
    age >= 10 & age < 20 ~ "adolescence",
    age >= 20 & age < 25 ~ "youth",
    age >= 25 & age < 65 ~ "adulthood",
    age >= 65 ~ "older adults"
  )) %>%
  
  
  # create ethnicity classes
  
  mutate(ethnicity = case_when(
    ethnicity == 0 ~ "African American",
    ethnicity == 1 ~ "Asian",
    ethnicity == 2 ~ "Caucasian",
    ethnicity == 3 ~ "Hispanic",
    ethnicity == 4 ~ "Native American",
    ethnicity == 5 ~ "Other/Unknown",
    TRUE ~ NA_character_
  ))


#-------------------------------------------------------------------------------

# PERFORMANCE MEASURES AT GLOBAL LEVEL

## sensitivity
sens <- sensitivity_vec(final_predictions$pred, final_predictions$obs,
                    negative = "Alive")

## specificity
spec <- specificity_vec(final_predictions$pred, final_predictions$obs,
                    positive = "Expire")

## precision
prec <- posPredValue(final_predictions$pred, final_predictions$obs,
                     negative = "Alive")

## recall
recall <- sensitivity_vec(final_predictions$pred, final_predictions$obs,
                    negative = "Alive")

## Global F1-measure

# f1_global <- F_meas(final_predictions$pred, final_predictions$obs) # Another way to obtain it
                    
F1_global <-F1_Score(final_predictions$obs, final_predictions$pred)

#-------------------------------------------------------------------------------

# CATEGORIES #


## sex ##

# create only males subset

final_subgroup <- final_predictions %>%
  filter(sex == "0")


# calculation of performance measures at subgroup levels
## precision

prec_male <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall

recall_male <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure

f1_male <- F_meas(final_subgroup$pred, final_subgroup$obs)

F1_global <-F1_Score(final_predictions$obs, final_predictions$pred)

#---

# create only females subset

final_subgroup <- final_predictions %>%
  filter(sex == "1")


# calculation of performance measures at subgroup levels
## precision

prec_female <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall

recall_female <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure

f1_female <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

### AGE GROUPS ### 

# create only ADOLESCENTS subset (there is no children)
final_subgroup <- final_predictions %>%
  filter(age_class == "adolescence")

# calculation of performance measures at subgroup levels
## precision
prec_adol <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_adol <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_adol <- F_meas(final_subgroup$pred, final_subgroup$obs)


#---

# create only YOUTH subset

final_subgroup <- final_predictions %>%
  filter(age_class == "youth")

# calculation of performance measures at subgroup levels
## precision
prec_youth <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_youth <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_youth <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

# create only ADULTS subset 

final_subgroup <- final_predictions %>%
  filter(age_class == "adulthood")

# calculation of performance measures at subgroup levels
## precision
prec_adult <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_adult <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_adult <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

# create only OLDER ADULTS subset 

final_subgroup <- final_predictions %>%
  filter(age_class == "older adults")

# calculation of performance measures at subgroup levels
## precision
prec_older <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_older <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_older <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

### ETHNICITY ###

# create only AFRICAN AMERICAN subset
final_subgroup <- final_predictions %>%
  filter(ethnicity == "African American")

# calculation of performance measures at subgroup levels
## precision
prec_afam <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_afam <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_afam <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

# create only ASIAN subset
final_subgroup <- final_predictions %>%
  filter(ethnicity == "Asian")

# calculation of performance measures at subgroup levels
## precision
prec_asian <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_asian <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_asian <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

# create only CAUCASIAN subset
final_subgroup <- final_predictions %>%
  filter(ethnicity == "Caucasian")

# calculation of performance measures at subgroup levels
## precision
prec_cau <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_cau <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_cau <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

# create only HISPANIC subset
final_subgroup <- final_predictions %>%
  filter(ethnicity == "Hispanic")

# calculation of performance measures at subgroup levels
## precision
prec_hisp <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_hisp <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_hisp <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

# create only NATIVE AMERICAN subset
final_subgroup <- final_predictions %>%
  filter(ethnicity == "Native American")

# calculation of performance measures at subgroup levels
## precision
prec_natam <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_natam <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_natam <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

# create only OTHER/UNKNOWN subset
final_subgroup <- final_predictions %>%
  filter(ethnicity == "Other/Unknown")

# calculation of performance measures at subgroup levels
## precision
prec_other <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_other <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_other <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

### TEACHING STATUS ###

# create only teaching status subset
final_subgroup <- final_predictions %>%
  filter(teachingstatus == "1")

# calculation of performance measures at subgroup levels
## precision
prec_teach <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_teach <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_teach <- F_meas(final_subgroup$pred, final_subgroup$obs)

#---

# create only no teaching status subset
final_subgroup <- final_predictions %>%
  filter(teachingstatus == "0")

# calculation of performance measures at subgroup levels
## precision
prec_nonteach <- posPredValue(final_subgroup$pred, final_subgroup$obs,
                     negative = "Alive")

## recall
recall_nonteach <- sensitivity_vec(final_subgroup$pred, final_subgroup$obs,
                    negative = "Alive")

## F1-measure
f1_nonteach <- F_meas(final_subgroup$pred, final_subgroup$obs)

#-------------------------------------------------------------------------------                    
                                      
### CREATE FINAL DATASETS WITH PERFORMANCE VALUES

# by categories
data_plot <- data.frame(class = c("Global score",
                     "Men",
                     "Women",
                     "Global score",
                     "Adolescents",
                     "Youth",
                     "Adults",
                     "Older adults",
                     "Global score",
                     "African American", 
                     "Asian",
                     "Caucasian",
                     "Hispanic",
                     "Native American",
                     "Other/Unknown",
                     "Global score",
                     "Teaching\nstatus",
                     "Not teaching\nstatus"),
           category = c(rep("Gender", 3),
                        rep("Age", 5), 
                        rep("Ethnicity", 7),
                        rep("Teaching status", 3)),
           value = c(F1_global, f1_male, f1_female,
                     F1_global,
                     f1_adol, f1_youth, f1_adult,
                     f1_older, 
                     F1_global, f1_afam, f1_asian,
                     f1_cau, f1_hisp,
                     f1_natam, f1_other,
                     F1_global, f1_teach, f1_nonteach),
           col = c(c(0, 1, 1, 
                     0, 1, 1, 1, 1,
                     0, 1, 1, 1, 1, 1, 1,
                     0, 1, 1)))

#-------------------------------------------------------------------------------

### PLOTS ###

# plot by race
plot_race <- data_plot %>%
  filter(category == "Ethnicity") %>%
  mutate(class = reorder(class, value)) %>%
  mutate(col = factor(col),
         class = factor(class, 
                        levels = c("Asian",
                                   "Native American",
                                   "Other/Unknown",
                                   "Caucasian",
                                   "African American",
                                   "Hispanic",
                                   "Global score"))) %>%
  ggplot(aes(x = value, y = class, fill = col)) +
  geom_col() +
  geom_vline(xintercept = F1_global, linetype = 2) +
  geom_text(aes(label = format(round(value, 2), nsmall = 2)), color = "white", hjust = 2) +
  labs(title = "Ethnicity",
       x = "",
       y = "") +
  theme_bw() +
  theme(legend.position = "none") +
  scale_fill_nejm()

#---

# plot by age group
plot_age <- data_plot %>%
  filter(category == "Age") %>%
  mutate(class = reorder(class, value)) %>%
  mutate(col = factor(col),
         class = factor(class, 
                        levels = c("Older adults",
                                   "Adults",
                                   "Youth",
                                   "Adolescents",
                                   "Global score"))) %>%
  ggplot(aes(x = value, y = class, fill = col)) +
  geom_col() +
  geom_vline(xintercept = F1_global, linetype = 2) +
  geom_text(aes(label = format(round(value, 2), nsmall = 2)), color = "white", hjust = 2) +
  labs(title = "Age",
       x = "",
       y = "") +
  theme_bw() +
  theme(legend.position = "none") +
  scale_fill_nejm()

#---

# plot by gender
plot_gender <- data_plot %>%
  filter(category == "Gender") %>%
  mutate(class = reorder(class, value)) %>%
  mutate(col = factor(col),
         class = factor(class, 
                        levels = c("Men",
                                   "Women",
                                   "Global score"))) %>%
  ggplot(aes(x = value, y = class, fill = col)) +
  geom_col() +
  geom_vline(xintercept = F1_global, linetype = 2) +
  geom_text(aes(label = format(round(value, 2), nsmall = 2)), color = "white", hjust = 2) +
  labs(title = "Gender",
       x = "",
       y = "") +
  theme_bw() +
  theme(legend.position = "none") +
  scale_fill_nejm()

#---

# plot by teaching status
plot_teach <- data_plot %>%
  filter(category == "Teaching status") %>%
  mutate(class = reorder(class, value)) %>%
  mutate(col = factor(col),
         class = factor(class, 
                        levels = c("Not teaching\nstatus",
                                   "Teaching\nstatus",
                                   "Global score"))) %>%
  ggplot(aes(x = value, y = class, fill = col)) +
  geom_col() +
  geom_vline(xintercept = F1_global, linetype = 2) +
  geom_text(aes(label = format(round(value, 2), nsmall = 2)), color = "white", hjust = 2) +
  labs(title = "Teaching status",
       x = "",
       y = "") +
  theme_bw() +
  theme(legend.position = "none") +
  scale_fill_nejm()

#---

# combine plots
comb <- ggarrange(plot_gender, plot_age, 
                  plot_race, plot_teach)

# annotate figure
annotate_figure(comb,
         top = "ML Stratified Fairness Analysis",
         bottom = "F1-score",
         left = "Subgroup")

```


## Option B: We filter raw data first, and then we train the model with this data (dividing in train and test).

SUBGROUP 1: Sex - male

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

final_df_male <- final_df_selected %>%
  filter(sex == 0) %>%
  select(-sex)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_male$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_male = final_df_male[idx_train, ]
data_test_male  = final_df_male[-idx_train, ]

# 2º FORMULA

final_df_male_without_ID <- final_df_male %>%
  select(-patientunitstayid)

data_train_male_without_ID <- data_train_male %>%
  select(-patientunitstayid)

outcome_model_male <- "icu_mortality_72hrs"                                   
exposures_model_male <- names(data_train_male_without_ID)                   
exposures_model_male <- exposures_model_male[-which(exposures_model_male == outcome_model_male)]   

# Model formula

outcome_and_exposure_male <- as.formula(paste(outcome_model_male, "~", paste(exposures_model_male, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_male <- train(outcome_and_exposure_male,       
                data = data_train_male_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_male, file = "best_male_model.Rda")
fit_best_male <- readRDS("best_male_model.Rda")

# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_male, data_test_male, prob_threshold=0.4){        
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_male, newdata=data_test_male, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_male$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_male$patientunitstayid,                
                                 trueStatus = data_test_male$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_male = getFinalPredictions(fit_best_male, data_test_male)
final_predictions_male <- final_predictions_best_male


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_male$obs <- as.factor(final_predictions_male$obs)

levels(final_predictions_male$obs)

# mortality_prediction_B
final_predictions_male$pred <- as.factor(final_predictions_male$pred)

levels(final_predictions_male$pred)

# confusion matrix

matrix_1 <- confusionMatrix(final_predictions_male$pred, final_predictions_male$obs, mode = "everything")


# PRECISION

m1_Precision <- matrix_1$table[1, 1] / sum(matrix_1$table[1, ])

# RECALL

m1_recall <- matrix_1$table[1, 1] / sum(matrix_1$table[, 1])

```


SUBGROUP 2: Sex - female

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

final_df_female <- final_df_selected %>%
  filter(sex == 1) %>%
  select(-sex)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_female$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_female = final_df_female[idx_train, ]
data_test_female  = final_df_female[-idx_train, ]

# 2º FORMULA

final_df_female_without_ID <- final_df_female %>%
  select(-patientunitstayid)

data_train_female_without_ID <- data_train_female %>%
  select(-patientunitstayid)

outcome_model_female <- "icu_mortality_72hrs"                                   
exposures_model_female <- names(data_train_female_without_ID)                   
exposures_model_female <- exposures_model_female[-which(exposures_model_female == outcome_model_female)]   

# Model formula

outcome_and_exposure_female <- as.formula(paste(outcome_model_female, "~", paste(exposures_model_female, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_female <- train(outcome_and_exposure_female,      
                data = data_train_female_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_female, file = "best_female_model.Rda")
fit_best_female <- readRDS("best_female_model.Rda")

# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_female, data_test_female, prob_threshold=0.4){        
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_female, newdata=data_test_female, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_female$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_female$patientunitstayid,                
                                 trueStatus = data_test_female$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_female = getFinalPredictions(fit_best_female, data_test_female)
final_predictions_female <- final_predictions_best_female


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_female$obs <- as.factor(final_predictions_female$obs)

levels(final_predictions_female$obs)

# mortality_prediction_B
final_predictions_female$pred <- as.factor(final_predictions_female$pred)

levels(final_predictions_female$pred)

# confusion matrix

matrix_2 <- confusionMatrix(final_predictions_female$pred, final_predictions_female$obs, mode = "everything")


# PRECISION

m2_Precision <- matrix_2$table[1, 1] / sum(matrix_2$table[1, ])

# RECALL

m2_recall <- matrix_2$table[1, 1] / sum(matrix_2$table[, 1])

```


SUBGROUP 3: age - 0-9 years   # NO DATA (min age = 17 years old)

SUBGROUP 4: age - 10-19 years

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

final_df_adolescence <- final_df_selected %>%
  filter(age > 9) %>%
  filter(age < 20)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_adolescence$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_adolescence = final_df_adolescence[idx_train, ]
data_test_adolescence  = final_df_adolescence[-idx_train, ]

# 2º FORMULA

final_df_adolescence_without_ID <- final_df_adolescence %>%
  select(-patientunitstayid)

data_train_adolescence_without_ID <- data_train_adolescence %>%
  select(-patientunitstayid)

outcome_model_adolescence <- "icu_mortality_72hrs"                                   
exposures_model_adolescence <- names(data_train_adolescence_without_ID)                   
exposures_model_adolescence <- exposures_model_adolescence[-which(exposures_model_adolescence == outcome_model_adolescence)]   

# Model formula

outcome_and_exposure_adolescence <- as.formula(paste(outcome_model_adolescence, "~", paste(exposures_model_adolescence, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_adolescence <- train(outcome_and_exposure_adolescence,  
                data = data_train_adolescence_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_adolescence, file = "best_adolescence_model.Rda")
fit_best_adolescence <- readRDS("best_adolescence_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_adolescence, data_test_adolescence, prob_threshold=0.4){  
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_adolescence, newdata=data_test_adolescence, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_adolescence$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_adolescence$patientunitstayid,                
                                 trueStatus = data_test_adolescence$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_adolescence = getFinalPredictions(fit_best_adolescence, data_test_adolescence)
final_predictions_adolescence <- final_predictions_best_adolescence


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_adolescence$obs <- as.factor(final_predictions_adolescence$obs)

levels(final_predictions_adolescence$obs)

# mortality_prediction_B
final_predictions_adolescence$pred <- as.factor(final_predictions_adolescence$pred)

levels(final_predictions_adolescence$pred)

# confusion matrix

matrix_4 <- confusionMatrix(final_predictions_adolescence$pred, final_predictions_adolescence$obs, mode = "everything")


# PRECISION

m4_Precision <- matrix_4$table[1, 1] / sum(matrix_4$table[1, ])

# RECALL

m4_recall <- matrix_4$table[1, 1] / sum(matrix_4$table[, 1])

```


SUBGROUP 5: age - 20-24 years

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 5: age - 20-24 years

final_df_youth <- final_df_selected %>%
  filter(age > 19) %>%
  filter(age < 25)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_youth$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_youth = final_df_youth[idx_train, ]
data_test_youth  = final_df_youth[-idx_train, ]

# 2º FORMULA

final_df_youth_without_ID <- final_df_youth %>%
  select(-patientunitstayid)

data_train_youth_without_ID <- data_train_youth %>%
  select(-patientunitstayid)

outcome_model_youth <- "icu_mortality_72hrs"                                   
exposures_model_youth <- names(data_train_youth_without_ID)                   
exposures_model_youth <- exposures_model_youth[-which(exposures_model_youth == outcome_model_youth)]   

# Model formula

outcome_and_exposure_youth <- as.formula(paste(outcome_model_youth, "~", paste(exposures_model_youth, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_youth <- train(outcome_and_exposure_youth,
                data = data_train_youth_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_youth, file = "best_youth_model.Rda")
fit_best_youth <- readRDS("best_youth_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_youth, data_test_youth, prob_threshold=0.4){               
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_youth, newdata=data_test_youth, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_youth$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_youth$patientunitstayid,                
                                 trueStatus = data_test_youth$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_youth = getFinalPredictions(fit_best_youth, data_test_youth)
final_predictions_youth <- final_predictions_best_youth


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_youth$obs <- as.factor(final_predictions_youth$obs)

levels(final_predictions_youth$obs)

# mortality_prediction_B
final_predictions_youth$pred <- as.factor(final_predictions_youth$pred)

levels(final_predictions_youth$pred)

# confusion matrix

matrix_5 <- confusionMatrix(final_predictions_youth$pred, final_predictions_youth$obs, mode = "everything")


# PRECISION

m5_Precision <- matrix_5$table[1, 1] / sum(matrix_5$table[1, ])

# RECALL

m5_recall <- matrix_5$table[1, 1] / sum(matrix_5$table[, 1])

```


SUBGROUP 6: age - 25-64 years

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 6: age - 25-64 years

final_df_adulthood <- final_df_selected %>%
  filter(age > 24) %>%
  filter(age < 65)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_adulthood$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_adulthood = final_df_adulthood[idx_train, ]
data_test_adulthood  = final_df_adulthood[-idx_train, ]

# 2º FORMULA

final_df_adulthood_without_ID <- final_df_adulthood %>%
  select(-patientunitstayid)

data_train_adulthood_without_ID <- data_train_adulthood %>%
  select(-patientunitstayid)

outcome_model_adulthood <- "icu_mortality_72hrs"                                   
exposures_model_adulthood <- names(data_train_adulthood_without_ID)                   
exposures_model_adulthood <- exposures_model_adulthood[-which(exposures_model_adulthood == outcome_model_adulthood)]   

# Model formula

outcome_and_exposure_adulthood <- as.formula(paste(outcome_model_adulthood, "~", paste(exposures_model_adulthood, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_adulthood <- train(outcome_and_exposure_adulthood,   
                data = data_train_adulthood_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_adulthood, file = "best_adulthood_model.Rda")
fit_best_adulthood <- readRDS("best_adulthood_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_adulthood, data_test_adulthood, prob_threshold=0.4){                
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_adulthood, newdata=data_test_adulthood, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_adulthood$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_adulthood$patientunitstayid,                
                                 trueStatus = data_test_adulthood$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_adulthood = getFinalPredictions(fit_best_adulthood, data_test_adulthood)
final_predictions_adulthood <- final_predictions_best_adulthood


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_adulthood$obs <- as.factor(final_predictions_adulthood$obs)

levels(final_predictions_adulthood$obs)

# mortality_prediction_B
final_predictions_adulthood$pred <- as.factor(final_predictions_adulthood$pred)

levels(final_predictions_adulthood$pred)

# confusion matrix

matrix_6 <- confusionMatrix(final_predictions_adulthood$pred, final_predictions_adulthood$obs, mode = "everything")


# PRECISION

m6_Precision <- matrix_6$table[1, 1] / sum(matrix_6$table[1, ])

# RECALL

m6_recall <- matrix_6$table[1, 1] / sum(matrix_6$table[, 1])

```


SUBGROUP 7: age - >65 years

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 7: age - >65 years

final_df_old_age <- final_df_selected %>%
  filter(age > 64)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_old_age$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_old_age = final_df_old_age[idx_train, ]
data_test_old_age  = final_df_old_age[-idx_train, ]

# 2º FORMULA

final_df_old_age_without_ID <- final_df_old_age %>%
  select(-patientunitstayid)

data_train_old_age_without_ID <- data_train_old_age %>%
  select(-patientunitstayid)

outcome_model_old_age <- "icu_mortality_72hrs"                                   
exposures_model_old_age <- names(data_train_old_age_without_ID)                   
exposures_model_old_age <- exposures_model_old_age[-which(exposures_model_old_age == outcome_model_old_age)]   

# Model formula

outcome_and_exposure_old_age <- as.formula(paste(outcome_model_old_age, "~", paste(exposures_model_old_age, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_old_age <- train(outcome_and_exposure_old_age,       
                data = data_train_old_age_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_old_age, file = "best_old_age_model.Rda")
fit_best_old_age <- readRDS("best_old_age_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_old_age, data_test_old_age, prob_threshold=0.5){   
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_old_age, newdata=data_test_old_age, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_old_age$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_old_age$patientunitstayid,                
                                 trueStatus = data_test_old_age$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_old_age = getFinalPredictions(fit_best_old_age, data_test_old_age)
final_predictions_old_age <- final_predictions_best_old_age


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_old_age$obs <- as.factor(final_predictions_old_age$obs)

levels(final_predictions_old_age$obs)

# mortality_prediction_B
final_predictions_old_age$pred <- as.factor(final_predictions_old_age$pred)

levels(final_predictions_old_age$pred)

# confusion matrix

matrix_7 <- confusionMatrix(final_predictions_old_age$pred, final_predictions_old_age$obs, mode = "everything")


# PRECISION

m7_Precision <- matrix_7$table[1, 1] / sum(matrix_7$table[1, ])

# RECALL

m7_recall <- matrix_7$table[1, 1] / sum(matrix_7$table[, 1])

```


SUBGROUP 8: teachingstatus - 1 (TRUE)

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 8: teachingstatus - 1 (TRUE)

final_df_teach_true <- final_df_selected %>%
  filter(teachingstatus == 1) %>%
  select(-teachingstatus)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_teach_true$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_teach_true = final_df_teach_true[idx_train, ]
data_test_teach_true  = final_df_teach_true[-idx_train, ]

# 2º FORMULA

final_df_teach_true_without_ID <- final_df_teach_true %>%
  select(-patientunitstayid)

data_train_teach_true_without_ID <- data_train_teach_true %>%
  select(-patientunitstayid)

outcome_model_teach_true <- "icu_mortality_72hrs"                                   
exposures_model_teach_true <- names(data_train_teach_true_without_ID)                   
exposures_model_teach_true <- exposures_model_teach_true[-which(exposures_model_teach_true == outcome_model_teach_true)]   

# Model formula

outcome_and_exposure_teach_true <- as.formula(paste(outcome_model_teach_true, "~", paste(exposures_model_teach_true, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_teach_true <- train(outcome_and_exposure_teach_true,      
                data = data_train_teach_true_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_teach_true, file = "best_teach_true_model.Rda")
fit_best_teach_true <- readRDS("best_teach_true_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_teach_true, data_test_teach_true, prob_threshold=0.4){      
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_teach_true, newdata=data_test_teach_true, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_teach_true$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_teach_true$patientunitstayid,                
                                 trueStatus = data_test_teach_true$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_teach_true = getFinalPredictions(fit_best_teach_true, data_test_teach_true)
final_predictions_teach_true <- final_predictions_best_teach_true


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_teach_true$obs <- as.factor(final_predictions_teach_true$obs)

levels(final_predictions_teach_true$obs)

# mortality_prediction_B
final_predictions_teach_true$pred <- as.factor(final_predictions_teach_true$pred)

levels(final_predictions_teach_true$pred)

# confusion matrix

matrix_8 <- confusionMatrix(final_predictions_teach_true$pred, final_predictions_teach_true$obs, mode = "everything")


# PRECISION

m8_Precision <- matrix_8$table[1, 1] / sum(matrix_8$table[1, ])

# RECALL

m8_recall <- matrix_8$table[1, 1] / sum(matrix_8$table[, 1])

```


SUBGROUP 9: teachingstatus - 0 (FALSE)

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 9: teachingstatus - 0 (FALSE)

final_df_teach_false <- final_df_selected %>%
  filter(teachingstatus == 0) %>%
  select(-teachingstatus)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_teach_false$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_teach_false = final_df_teach_false[idx_train, ]
data_test_teach_false  = final_df_teach_false[-idx_train, ]

# 2º FORMULA

final_df_teach_false_without_ID <- final_df_teach_false %>%
  select(-patientunitstayid)

data_train_teach_false_without_ID <- data_train_teach_false %>%
  select(-patientunitstayid)

outcome_model_teach_false <- "icu_mortality_72hrs"                                   
exposures_model_teach_false <- names(data_train_teach_false_without_ID)                   
exposures_model_teach_false <- exposures_model_teach_false[-which(exposures_model_teach_false == outcome_model_teach_false)]   

# Model formula

outcome_and_exposure_teach_false <- as.formula(paste(outcome_model_teach_false, "~", paste(exposures_model_teach_false, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_teach_false <- train(outcome_and_exposure_teach_false,  
                data = data_train_teach_false_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_teach_false, file = "best_teach_false_model.Rda")
fit_best_teach_false <- readRDS("best_teach_false_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_teach_false, data_test_teach_false, prob_threshold=0.4){ 
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_teach_false, newdata=data_test_teach_false, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_teach_false$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_teach_false$patientunitstayid,                
                                 trueStatus = data_test_teach_false$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_teach_false = getFinalPredictions(fit_best_teach_false, data_test_teach_false)
final_predictions_teach_false <- final_predictions_best_teach_false


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_teach_false$obs <- as.factor(final_predictions_teach_false$obs)

levels(final_predictions_teach_false$obs)

# mortality_prediction_B
final_predictions_teach_false$pred <- as.factor(final_predictions_teach_false$pred)

levels(final_predictions_teach_false$pred)

# confusion matrix

matrix_9 <- confusionMatrix(final_predictions_teach_false$pred, final_predictions_teach_false$obs, mode = "everything")


# PRECISION

m9_Precision <- matrix_9$table[1, 1] / sum(matrix_9$table[1, ])

# RECALL

m9_recall <- matrix_9$table[1, 1] / sum(matrix_9$table[, 1])

```


SUBGROUP 10: ethnicity - 0 (African American)

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 10: ethnicity - 0 (African American)

final_df_african_american <- final_df_selected %>%
  filter(ethnicity == 0) %>%
  select(-ethnicity)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_african_american$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_african_american = final_df_african_american[idx_train, ]
data_test_african_american  = final_df_african_american[-idx_train, ]

# 2º FORMULA

final_df_african_american_without_ID <- final_df_african_american %>%
  select(-patientunitstayid)

data_train_african_american_without_ID <- data_train_african_american %>%
  select(-patientunitstayid)

outcome_model_african_american <- "icu_mortality_72hrs"                                   
exposures_model_african_american <- names(data_train_african_american_without_ID)                   
exposures_model_african_american <- exposures_model_african_american[-which(exposures_model_african_american == outcome_model_african_american)]   

# Model formula

outcome_and_exposure_african_american <- as.formula(paste(outcome_model_african_american, "~", paste(exposures_model_african_american, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_african_american <- train(outcome_and_exposure_african_american,    
                data = data_train_african_american_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_african_american, file = "best_african_american_model.Rda")
fit_best_african_american <- readRDS("best_african_american_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_african_american, data_test_african_american, prob_threshold=0.4){ 
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_african_american, newdata=data_test_african_american, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_african_american$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_african_american$patientunitstayid,                
                                 trueStatus = data_test_african_american$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_african_american = getFinalPredictions(fit_best_african_american, data_test_african_american)
final_predictions_african_american <- final_predictions_best_african_american


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_african_american$obs <- as.factor(final_predictions_african_american$obs)

levels(final_predictions_african_american$obs)

# mortality_prediction_B
final_predictions_african_american$pred <- as.factor(final_predictions_african_american$pred)

levels(final_predictions_african_american$pred)

# confusion matrix

matrix_10 <- confusionMatrix(final_predictions_african_american$pred, final_predictions_african_american$obs, mode = "everything")


# PRECISION

m10_Precision <- matrix_10$table[1, 1] / sum(matrix_10$table[1, ])

# RECALL

m10_recall <- matrix_10$table[1, 1] / sum(matrix_10$table[, 1])

```


SUBGROUP 11: ethnicity - 1 (Asian)

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 11: ethnicity - 1 (Asian)

final_df_asian <- final_df_selected %>%
  filter(ethnicity == 1) %>%
  select(-ethnicity)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_asian$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_asian = final_df_asian[idx_train, ]
data_test_asian  = final_df_asian[-idx_train, ]

# 2º FORMULA

final_df_asian_without_ID <- final_df_asian %>%
  select(-patientunitstayid)

data_train_asian_without_ID <- data_train_asian %>%
  select(-patientunitstayid)

outcome_model_asian <- "icu_mortality_72hrs"                                   
exposures_model_asian <- names(data_train_asian_without_ID)                   
exposures_model_asian <- exposures_model_asian[-which(exposures_model_asian == outcome_model_asian)]   

# Model formula

outcome_and_exposure_asian <- as.formula(paste(outcome_model_asian, "~", paste(exposures_model_asian, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_asian <- train(outcome_and_exposure_asian,    
                data = data_train_asian_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_asian, file = "best_asian_model.Rda")
fit_best_asian <- readRDS("best_asian_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_asian, data_test_asian, prob_threshold=0.4){     
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_asian, newdata=data_test_asian, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_asian$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_asian$patientunitstayid,                
                                 trueStatus = data_test_asian$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_asian = getFinalPredictions(fit_best_asian, data_test_asian)
final_predictions_asian <- final_predictions_best_asian


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_asian$obs <- as.factor(final_predictions_asian$obs)

levels(final_predictions_asian$obs)

# mortality_prediction_B
final_predictions_asian$pred <- as.factor(final_predictions_asian$pred)

levels(final_predictions_asian$pred)

# confusion matrix

matrix_11 <- confusionMatrix(final_predictions_asian$pred, final_predictions_asian$obs, mode = "everything")


# PRECISION

m11_Precision <- matrix_11$table[1, 1] / sum(matrix_11$table[1, ])

# RECALL

m11_recall <- matrix_11$table[1, 1] / sum(matrix_11$table[, 1])

```


SUBGROUP 12: ethnicity - 2 (Caucasian)

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 12: ethnicity - 2 (Caucasian)

final_df_caucasian <- final_df_selected %>%
  filter(ethnicity == 2) %>%
  select(-ethnicity)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_caucasian$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_caucasian = final_df_caucasian[idx_train, ]
data_test_caucasian  = final_df_caucasian[-idx_train, ]

# 2º FORMULA

final_df_caucasian_without_ID <- final_df_caucasian %>%
  select(-patientunitstayid)

data_train_caucasian_without_ID <- data_train_caucasian %>%
  select(-patientunitstayid)

outcome_model_caucasian <- "icu_mortality_72hrs"                                   
exposures_model_caucasian <- names(data_train_caucasian_without_ID)                   
exposures_model_caucasian <- exposures_model_caucasian[-which(exposures_model_caucasian == outcome_model_caucasian)]   

# Model formula

outcome_and_exposure_caucasian <- as.formula(paste(outcome_model_caucasian, "~", paste(exposures_model_caucasian, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_caucasian <- train(outcome_and_exposure_caucasian,   
                data = data_train_caucasian_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_caucasian, file = "best_caucasian_model.Rda")
fit_best_caucasian <- readRDS("best_caucasian_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_caucasian, data_test_caucasian, prob_threshold=0.4){      
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_caucasian, newdata=data_test_caucasian, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_caucasian$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_caucasian$patientunitstayid,                
                                 trueStatus = data_test_caucasian$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_caucasian = getFinalPredictions(fit_best_caucasian, data_test_caucasian)
final_predictions_caucasian <- final_predictions_best_caucasian


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_caucasian$obs <- as.factor(final_predictions_caucasian$obs)

levels(final_predictions_caucasian$obs)

# mortality_prediction_B
final_predictions_caucasian$pred <- as.factor(final_predictions_caucasian$pred)

levels(final_predictions_caucasian$pred)

# confusion matrix

matrix_12 <- confusionMatrix(final_predictions_caucasian$pred, final_predictions_caucasian$obs, mode = "everything")


# PRECISION

m12_Precision <- matrix_12$table[1, 1] / sum(matrix_12$table[1, ])

# RECALL

m12_recall <- matrix_12$table[1, 1] / sum(matrix_12$table[, 1])

```


SUBGROUP 13: ethnicity - 3 (Hispanic)

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 13: ethnicity - 3 (Hispanic)

final_df_hispanic <- final_df_selected %>%
  filter(ethnicity == 3) %>%
  select(-ethnicity)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_hispanic$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_hispanic = final_df_hispanic[idx_train, ]
data_test_hispanic  = final_df_hispanic[-idx_train, ]

# 2º FORMULA

final_df_hispanic_without_ID <- final_df_hispanic %>%
  select(-patientunitstayid)

data_train_hispanic_without_ID <- data_train_hispanic %>%
  select(-patientunitstayid)

outcome_model_hispanic <- "icu_mortality_72hrs"                                   
exposures_model_hispanic <- names(data_train_hispanic_without_ID)                   
exposures_model_hispanic <- exposures_model_hispanic[-which(exposures_model_hispanic == outcome_model_hispanic)]   

# Model formula

outcome_and_exposure_hispanic <- as.formula(paste(outcome_model_hispanic, "~", paste(exposures_model_hispanic, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_hispanic <- train(outcome_and_exposure_hispanic,       
                data = data_train_hispanic_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_hispanic, file = "best_hispanic_model.Rda")
fit_best_hispanic <- readRDS("best_hispanic_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_hispanic, data_test_hispanic, prob_threshold=0.4){      
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_hispanic, newdata=data_test_hispanic, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_hispanic$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_hispanic$patientunitstayid,                
                                 trueStatus = data_test_hispanic$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_hispanic = getFinalPredictions(fit_best_hispanic, data_test_hispanic)
final_predictions_hispanic <- final_predictions_best_hispanic


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_hispanic$obs <- as.factor(final_predictions_hispanic$obs)

levels(final_predictions_hispanic$obs)

# mortality_prediction_B
final_predictions_hispanic$pred <- as.factor(final_predictions_hispanic$pred)

levels(final_predictions_hispanic$pred)

# confusion matrix

matrix_13 <- confusionMatrix(final_predictions_hispanic$pred, final_predictions_hispanic$obs, mode = "everything")


# PRECISION

m13_Precision <- matrix_13$table[1, 1] / sum(matrix_13$table[1, ])

# RECALL

m13_recall <- matrix_13$table[1, 1] / sum(matrix_13$table[, 1])

```


SUBGROUP 14: ethnicity - 4 (Native American)

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 14: ethnicity - 4 (Native American)

final_df_native_american <- final_df_selected %>%
  filter(ethnicity == 4) %>%
  select(-ethnicity)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_native_american$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_native_american = final_df_native_american[idx_train, ]
data_test_native_american  = final_df_native_american[-idx_train, ]

# 2º FORMULA

final_df_native_american_without_ID <- final_df_native_american %>%
  select(-patientunitstayid)

data_train_native_american_without_ID <- data_train_native_american %>%
  select(-patientunitstayid)

outcome_model_native_american <- "icu_mortality_72hrs"                                   
exposures_model_native_american <- names(data_train_native_american_without_ID)                   
exposures_model_native_american <- exposures_model_native_american[-which(exposures_model_native_american == outcome_model_native_american)]   

# Model formula

outcome_and_exposure_native_american <- as.formula(paste(outcome_model_native_american, "~", paste(exposures_model_native_american, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_native_american <- train(outcome_and_exposure_native_american,     
                data = data_train_native_american_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_native_american, file = "best_native_american_model.Rda")
fit_best_native_american <- readRDS("best_native_american_model.Rda")


# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_native_american, data_test_native_american, prob_threshold=0.4){  
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_native_american, newdata=data_test_native_american, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_native_american$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_native_american$patientunitstayid,                
                                 trueStatus = data_test_native_american$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_native_american = getFinalPredictions(fit_best_native_american, data_test_native_american)
final_predictions_native_american <- final_predictions_best_native_american


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_native_american$obs <- as.factor(final_predictions_native_american$obs)

levels(final_predictions_native_american$obs)

# mortality_prediction_B
final_predictions_native_american$pred <- as.factor(final_predictions_native_american$pred)

levels(final_predictions_native_american$pred)

# confusion matrix

matrix_14 <- confusionMatrix(final_predictions_native_american$pred, final_predictions_native_american$obs, mode = "everything")


# PRECISION

m14_Precision <- matrix_14$table[1, 1] / sum(matrix_14$table[1, ])

# RECALL

m14_recall <- matrix_14$table[1, 1] / sum(matrix_14$table[, 1])

```


SUBGROUP 15: ethnicity - 5 (Other/Unknown)

```{r}

# Set the seed to make our partition reproducible.
set.seed(123)

#-------------------------------------------------------------------------------

# SUBGROUP 15: ethnicity - 5 (Other/Unknown)

final_df_ethnicity_other <- final_df_selected %>%
  filter(ethnicity == 5) %>%
  select(-ethnicity)

# 1. TRAIN-TEST SELECTION

# Select the target column.
idx_train  = createDataPartition(as.factor(final_df_ethnicity_other$icu_mortality_72hrs), times = 1, p = 0.80, list=F)[,1]

# Division into train and test
data_train_ethnicity_other = final_df_ethnicity_other[idx_train, ]
data_test_ethnicity_other  = final_df_ethnicity_other[-idx_train, ]

# 2º FORMULA

final_df_ethnicity_other_without_ID <- final_df_ethnicity_other %>%
  select(-patientunitstayid)

data_train_ethnicity_other_without_ID <- data_train_ethnicity_other %>%
  select(-patientunitstayid)

outcome_model_ethnicity_other <- "icu_mortality_72hrs"                                   
exposures_model_ethnicity_other <- names(data_train_ethnicity_other_without_ID)                   
exposures_model_ethnicity_other <- exposures_model_ethnicity_other[-which(exposures_model_ethnicity_other == outcome_model_ethnicity_other)]   

# Model formula

outcome_and_exposure_ethnicity_other <- as.formula(paste(outcome_model_ethnicity_other, "~", paste(exposures_model_ethnicity_other, collapse = " + "))) 

# 3. TRAINING THE MODEL

fit_best_ethnicity_other <- train(outcome_and_exposure_ethnicity_other,   
                data = data_train_ethnicity_other_without_ID %>% as.data.frame(),
                method = "rf",
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")

saveRDS(fit_best_ethnicity_other, file = "best_ethnicity_other_model.Rda")
fit_best_ethnicity_other <- readRDS("best_ethnicity_other_model.Rda")

# 4. FINAL PREDICTIONS

getFinalPredictions = function(fit_best_ethnicity_other, data_test_ethnicity_other, prob_threshold=0.4){   
                                                                                           
  prediction_probabilities_tmp = predict(fit_best_ethnicity_other, newdata=data_test_ethnicity_other, type = "prob")        

  rownames(prediction_probabilities_tmp) = data_test_ethnicity_other$patientunitstayid          

  # Match the prediction with current data.
  
  final_predictions_tmp = tibble(ID         = data_test_ethnicity_other$patientunitstayid,                
                                 trueStatus = data_test_ethnicity_other$icu_mortality_72hrs,               # Change target variable
                                 obs        = if_else(trueStatus == "Expired",1,0))        # If Expired, enter 1.

  final_predictions_tmp = cbind(final_predictions_tmp,                                    
                                prediction_probabilities_tmp[match(final_predictions_tmp$ID,    # Match using ID.
                                                                   rownames(prediction_probabilities_tmp)),])
  
  final_predictions_tmp = final_predictions_tmp %>%
    dplyr::mutate(pred=dplyr::if_else(Alive > prob_threshold, 0, 1))
  
  final_predictions_tmp
}

# End of function

final_predictions_best_ethnicity_other = getFinalPredictions(fit_best_ethnicity_other, data_test_ethnicity_other)
final_predictions_ethnicity_other <- final_predictions_best_ethnicity_other


# 5. CONFUSION MATRIX


# icu_mortality_72hrs
final_predictions_ethnicity_other$obs <- as.factor(final_predictions_ethnicity_other$obs)

levels(final_predictions_ethnicity_other$obs)

# mortality_prediction_B
final_predictions_ethnicity_other$pred <- as.factor(final_predictions_ethnicity_other$pred)

levels(final_predictions_ethnicity_other$pred)

# confusion matrix

matrix_15 <- confusionMatrix(final_predictions_ethnicity_other$pred, final_predictions_ethnicity_other$obs, mode = "everything")


# PRECISION

m15_Precision <- matrix_15$table[1, 1] / sum(matrix_15$table[1, ])

# RECALL

m15_recall <- matrix_15$table[1, 1] / sum(matrix_15$table[, 1])

```


## Comparative table Option B

```{r}

# UPLOAD THE MODELS (Not needed if you have run all the code)

fit_best_male <- readRDS("best_male_model.Rda")          
fit_best_female <- readRDS("best_female_model.Rda")          
fit_best_adolescence <- readRDS("best_adolescence_model.Rda")             
fit_best_youth <- readRDS("best_youth_model.Rda")                         
fit_best_adulthood <- readRDS("best_adulthood_model.Rda")                 
fit_best_old_age <- readRDS("best_old_age_model.Rda")
fit_best_teach_true <- readRDS("best_teach_true_model.Rda")
fit_best_teach_false <- readRDS("best_teach_false_model.Rda")
fit_best_african_american <- readRDS("best_african_american_model.Rda")
fit_best_asian <- readRDS("best_asian_model.Rda")
fit_best_caucasian <- readRDS("best_caucasian_model.Rda")
fit_best_hispanic <- readRDS("best_hispanic_model.Rda")
fit_best_native_american <- readRDS("best_native_american_model.Rda")    
fit_best_ethnicity_other <- readRDS("best_ethnicity_other_model.Rda")

# Before executing the following code it is necessary to execute the prediction section of each SUBGROUP.

# PARAMETER F1.

F1_global <-F1_Score(final_predictions$obs, final_predictions$pred)

best_F1_male <-F1_Score(final_predictions_best_male$obs, final_predictions_best_male$pred)
best_F1_female <-F1_Score(final_predictions_best_female$obs, final_predictions_best_female$pred)
best_F1_adolescence <-F1_Score(final_predictions_best_adolescence$obs, final_predictions_best_adolescence$pred)
best_F1_youth <-F1_Score(final_predictions_best_youth$obs, final_predictions_best_youth$pred)
best_F1_adulthood<-F1_Score(final_predictions_best_adulthood$obs, final_predictions_best_adulthood$pred)
best_F1_old_age <-F1_Score(final_predictions_best_old_age$obs, final_predictions_best_old_age$pred)
best_F1_teach_true <-F1_Score(final_predictions_best_teach_true$obs, final_predictions_best_teach_true$pred)
best_F1_teach_false <-F1_Score(final_predictions_best_teach_false$obs, final_predictions_best_teach_false$pred)
best_F1_african_american <-F1_Score(final_predictions_best_african_american$obs, final_predictions_best_african_american$pred)
best_F1_asian <-F1_Score(final_predictions_best_asian$obs, final_predictions_best_asian$pred)
best_F1_caucasian <-F1_Score(final_predictions_best_caucasian$obs, final_predictions_best_caucasian$pred)
best_F1_hispanic <-F1_Score(final_predictions_best_hispanic$obs, final_predictions_best_hispanic$pred)
best_F1_native_american <-F1_Score(final_predictions_best_native_american$obs, final_predictions_best_native_american$pred)
best_F1_ethnicity_other <-F1_Score(final_predictions_best_ethnicity_other$obs, final_predictions_best_ethnicity_other$pred)

#-------------------------------------------------------------------------------

# CREATE THE TABLE

# Models
Groups <- c("Male", "Female", "Adolescence", "Youth", "Adulthood", "Old Age", "Teach True", "Teach False", "African American", "Asian", "Caucasian", "Hispanic", "Native American", "Ethnicity Other")

# Precision
Precision <- round(c(m1_Precision, m2_Precision, m4_Precision, m5_Precision, m6_Precision, m7_Precision, m8_Precision, m9_Precision, m10_Precision, m11_Precision, m12_Precision, m13_Precision, m14_Precision, m15_Precision),3)

# Recall
Recall <- round(c(m1_recall, m2_recall, m4_recall, m5_recall, m6_recall, m7_recall, m8_recall, m9_recall, m10_recall, m11_recall, m12_recall, m13_recall, m14_recall, m15_recall),3)

# F1 parameter
F1 <- round(c(best_F1_male, best_F1_female, best_F1_adolescence, best_F1_youth, best_F1_adulthood, best_F1_old_age, best_F1_teach_true, best_F1_teach_false, best_F1_african_american, best_F1_asian, best_F1_caucasian, best_F1_hispanic, best_F1_native_american, best_F1_ethnicity_other),3)

# Dataframe
comparative_option_B <- data.frame(Groups = Groups, Precision = Precision, Recall = Recall, F1 = F1)

print(comparative_option_B)

```


# PLOT Option B

```{r}
 
# CREATE FINAL DATASETS WITH PERFORMANCE VALUES (by categories)

data_plot <- data.frame(class = c("Global score",           
                     "Men",
                     "Women",
                     "Global score",
                     "Adolescence",
                     "Youth",
                     "Adults",
                     "Older adults",
                     "Global score",
                     "African American", 
                     "Asian",
                     "Caucasian",
                     "Hispanic",
                     "Native American",
                     "Other/Unknown",
                     "Global score",
                     "Teaching\nstatus",
                     "Not teaching\nstatus"),
           category = c(rep("Gender", 3),
                        rep("Age", 5), 
                        rep("Ethnicity", 7),
                        rep("Teaching status", 3)),
           value = c(F1_global, best_F1_male, best_F1_female,
                     F1_global,
                     best_F1_adolescence,
                     best_F1_youth,
                     best_F1_adulthood,
                     best_F1_old_age, 
                     F1_global, best_F1_african_american, best_F1_asian,
                     best_F1_caucasian, best_F1_hispanic,
                     best_F1_native_american, best_F1_ethnicity_other,
                     F1_global, best_F1_teach_true, best_F1_teach_false),
           col = c(c(0, 1, 1, 
                     0, 1, 1, 1, 1,
                     0, 1, 1, 1, 1, 1, 1,
                     0, 1, 1)))


#---

# Plot by race                                                       

plot_race <- data_plot %>%
  filter(category == "Ethnicity") %>%
  mutate(class = reorder(class, value)) %>%
  mutate(col = factor(col),
         class = factor(class, 
                        levels = c("Asian",
                                   "Native American",
                                   "Other/Unknown",
                                   "Caucasian",
                                   "African American",
                                   "Hispanic",
                                   "Global score"))) %>%
  ggplot(aes(x = value, y = class, fill = col)) +
  geom_col() +
  geom_vline(xintercept = F1_global, linetype = 2) +
  geom_text(aes(label = format(round(value, 2), nsmall = 2)), color = "white", hjust = 2) +
  labs(title = "Ethnicity",
       x = "",
       y = "") +
  theme_bw() +
  theme(legend.position = "none") +
  scale_fill_nejm()

#---

# Plot by age group                                              
plot_age <- data_plot %>%
  filter(category == "Age") %>%
  mutate(class = reorder(class, value)) %>%
  mutate(col = factor(col),
         class = factor(class, 
                        levels = c("Older adults",
                                   "Adults",
                                   "Adolescence",
                                   "Youth",
                                   "Global score"))) %>%
  ggplot(aes(x = value, y = class, fill = col)) +
  geom_col() +
  geom_vline(xintercept = F1_global, linetype = 2) +
  geom_text(aes(label = format(round(value, 2), nsmall = 2)), color = "white", hjust = 2) +
  labs(title = "Age",
       x = "",
       y = "") +
  theme_bw() +
  theme(legend.position = "none") +
  scale_fill_nejm()

#---

# Plot by gender

plot_gender <- data_plot %>%
  filter(category == "Gender") %>%
  mutate(class = reorder(class, value)) %>%
  mutate(col = factor(col),
         class = factor(class, 
                        levels = c("Men",
                                   "Women",
                                   "Global score"))) %>%
  ggplot(aes(x = value, y = class, fill = col)) +
  geom_col() +
  geom_vline(xintercept = F1_global, linetype = 2) +
  geom_text(aes(label = format(round(value, 2), nsmall = 2)), color = "white", hjust = 2) +
  labs(title = "Gender",
       x = "",
       y = "") +
  theme_bw() +
  theme(legend.position = "none") +
  scale_fill_nejm()

#---

# Plot by teaching status

plot_teach <- data_plot %>%
  filter(category == "Teaching status") %>%
  mutate(class = reorder(class, value)) %>%
  mutate(col = factor(col),
         class = factor(class, 
                        levels = c("Not teaching\nstatus",
                                   "Teaching\nstatus",
                                   "Global score"))) %>%
  ggplot(aes(x = value, y = class, fill = col)) +
  geom_col() +
  geom_vline(xintercept = F1_global, linetype = 2) +
  geom_text(aes(label = format(round(value, 2), nsmall = 2)), color = "white", hjust = 2) +
  labs(title = "Teaching status",
       x = "",
       y = "") +
  theme_bw() +
  theme(legend.position = "none") +
  scale_fill_nejm()


# Combine plots

comb <- ggarrange(plot_gender, plot_age, 
                  plot_race, plot_teach)

# Annotate figure

annotate_figure(comb,
         top = "ML Stratified Fairness Analysis",
         bottom = "F1-score",
         left = "Subgroup")

```



# OUTCOMES: PREDICTIONS USING `data_test`

```{r}

# 1. gbm

##BOOTSTRAPPING SAMPLING

# Aimed to predict a 95% confidence interval for predictions, we are going to conduct our model applying bootstrapping sampling (10 times)
# bootstrapping

set.seed(42)

# number of resamplings
k <- 10

## creation of a empty matrix for predictions
predictions <- matrix(nrow = nrow(data_test), ncol = k)

# loop for computing predictions for each sampling process
for (i in 1:k) {
  sample_indices <- sample(nrow(data_train), replace = TRUE)
  bootstrap_sample <- data_train[sample_indices, ]
  fit_best_new <- train(outcome_and_exposure,
                data = bootstrap_sample,
                method = "gbm",                                                  # Change the model of interest
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")
  # Predicting probability of class 1
  predictions[, i] <- getFinalPredictions(fit_best_new, data_test)$Expired
}


# SAVE PREDICTIONS
saveRDS(predictions, file = "pred_folds_gbm_C.Rda")

#-------------------------------------------------------------------------------


# 2. rf

##BOOTSTRAPPING SAMPLING

# Aimed to predict a 95% confidence interval for predictions, we are going to conduct our model applying bootstrapping sampling (10 times)
# bootstrapping

set.seed(42)

# number of resamplings
k <- 10

## creation of a empty matrix for predictions
predictions <- matrix(nrow = nrow(data_test), ncol = k)

# loop for computing predictions for each sampling process
for (i in 1:k) {
  sample_indices <- sample(nrow(data_train), replace = TRUE)
  bootstrap_sample <- data_train[sample_indices, ]
  fit_best_new <- train(outcome_and_exposure,
                data = bootstrap_sample,
                method = "rf",                                                  # Change the model of interest
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")
  # Predicting probability of class 1
  predictions[, i] <- getFinalPredictions(fit_best_new, data_test)$Expired
}


# SAVE PREDICTIONS
saveRDS(predictions, file = "pred_folds_rf_C.Rda")

#-------------------------------------------------------------------------------


# 3. xgb

##BOOTSTRAPPING SAMPLING

# Aimed to predict a 95% confidence interval for predictions, we are going to conduct our model applying bootstrapping sampling (10 times)
# bootstrapping

set.seed(42)

# number of resamplings
k <- 10

## creation of a empty matrix for predictions
predictions <- matrix(nrow = nrow(data_test), ncol = k)

# loop for computing predictions for each sampling process
for (i in 1:k) {
  sample_indices <- sample(nrow(data_train), replace = TRUE)
  bootstrap_sample <- data_train[sample_indices, ]
  fit_best_new <- train(outcome_and_exposure,
                data = bootstrap_sample,
                method = "xgb",                                                  # Change the model of interest
                trControl = fitControl,
                verbose = F,
                na.action = na.omit,
                metric = "F1")
  # Predicting probability of class 1
  predictions[, i] <- getFinalPredictions(fit_best_new, data_test)$Expired
}


# SAVE PREDICTIONS
saveRDS(predictions, file = "pred_folds_xgb_C.Rda")

#-------------------------------------------------------------------------------

# LOAD PREDICTIONS
predictions_gbm <- readRDS("pred_folds_gbm_C.Rda")
predictions_rf <- readRDS("pred_folds_rf_C.Rda")
predictions_xgb <- readRDS("pred_folds_xgb_C.Rda")

#-------------------------------------------------------------------------------

# DATA TREATMENT

# Median for each row in prediction matrix
newdata_gbm <- apply(predictions_gbm, 1, median_qi)
# Plot!
data_plot_gbm <- do.call(rbind, newdata_gbm)

#---

# Median for each row in prediction matrix
newdata_rf <- apply(predictions_rf, 1, median_qi)
# Plot!
data_plot_rf <- do.call(rbind, newdata_rf)

#---

# Median for each row in prediction matrix
newdata_xgb <- apply(predictions_xgb, 1, median_qi)
# Plot!
data_plot_xgb <- do.call(rbind, newdata_xgb)

```


## prev_dept_los_hrs

```{r}

# 1. gbm

# plot regression line prev_dept_los_hrs
data_test %>%
  mutate(pred = data_plot_gbm$y,
         pred.lower = data_plot_gbm$ymin,
         pred.upper = data_plot_gbm$ymax) %>%
  filter(prev_dept_los_hrs < 100) %>%
  ggplot(aes(x = prev_dept_los_hrs, y = pred)) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred = ifelse(icu_mortality_72hrs == "Expired", 1, 0)) %>%
               filter(prev_dept_los_hrs <= 100) %>%
  filter(prev_rounded %in% seq(0, 100, 1))) +
  geom_point(alpha = 0.04) +
  geom_smooth(fill = "grey50",
              col = "black") +
  labs(title = "gbm",
       x = "Hours from general admission to ICU admission",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(0, 100, 5)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

#---

# 2. rf

# plot regression line prev_dept_los_hrs
data_test %>%
  mutate(pred = data_plot_rf$y,
         pred.lower = data_plot_rf$ymin,
         pred.upper = data_plot_rf$ymax) %>%
  filter(prev_dept_los_hrs < 100) %>%
  ggplot(aes(x = prev_dept_los_hrs, y = pred)) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred = ifelse(icu_mortality_72hrs == "Expired", 1, 0)) %>%
               filter(prev_dept_los_hrs <= 100) %>%
  filter(prev_rounded %in% seq(0, 100, 1))) +
  geom_point(alpha = 0.04) +
  geom_smooth(fill = "grey50",
              col = "black") +
  labs(title = "rf",
       x = "Hours from general admission to ICU admission",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(0, 100, 5)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

#---

# 3. xgb

# plot regression line prev_dept_los_hrs
data_test %>%
  mutate(pred = data_plot_xgb$y,
         pred.lower = data_plot_xgb$ymin,
         pred.upper = data_plot_xgb$ymax) %>%
  filter(prev_dept_los_hrs < 100) %>%
  ggplot(aes(x = prev_dept_los_hrs, y = pred)) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred = ifelse(icu_mortality_72hrs == "Expired", 1, 0)) %>%
               filter(prev_dept_los_hrs <= 100) %>%
  filter(prev_rounded %in% seq(0, 100, 1))) +
  geom_point(alpha = 0.04) +
  geom_smooth(fill = "grey50",
              col = "black") +
  labs(title = "xgb",
       x = "Hours from general admission to ICU admission",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(0, 100, 5)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))


```


## age

```{r}

# 1. gbm

# plot regression line age
data_test %>%
  mutate(pred = data_plot_gbm$y,
         pred.lower = data_plot_gbm$ymin,
         pred.upper = data_plot_gbm$ymax) %>%
  ggplot(aes(x = age, y = pred)) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred = ifelse(icu_mortality_72hrs == "Expired", 1, 0))) +
  geom_point(alpha = 0.04) +
  geom_smooth(fill = "grey50",
              col = "black") +
  labs(title= "gbm",
       x = "Age",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(0, 100, 5)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

#---

# 2. rf

# plot regression line age
data_test %>%
  mutate(pred = data_plot_rf$y,
         pred.lower = data_plot_rf$ymin,
         pred.upper = data_plot_rf$ymax) %>%
  ggplot(aes(x = age, y = pred)) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred = ifelse(icu_mortality_72hrs == "Expired", 1, 0))) +
  geom_point(alpha = 0.04) +
  geom_smooth(fill = "grey50",
              col = "black") +
  labs(title= "rf",
       x = "Age",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(0, 100, 5)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

#---

# 3. xgb

# plot regression line age
data_test %>%
  mutate(pred = data_plot_xgb$y,
         pred.lower = data_plot_xgb$ymin,
         pred.upper = data_plot_xgb$ymax) %>%
  ggplot(aes(x = age, y = pred)) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred = ifelse(icu_mortality_72hrs == "Expired", 1, 0))) +
  geom_point(alpha = 0.04) +
  geom_smooth(fill = "grey50",
              col = "black") +
  labs(title= "xgb",
       x = "Age",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(0, 100, 5)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

```


## norepinephrine

```{r}

# 1. gbm


# norepinephrine plots
data_test %>%
  mutate(pred = data_plot_gbm$y,
         pred.lower = data_plot_gbm$ymin,
         pred.upper = data_plot_gbm$ymax) %>%
  dplyr::group_by(norepinephrine) %>%
  dplyr::summarise(pred.mean = mean(pred),
            pred.low = mean(pred.lower),
            pred.up = mean(pred.upper)) %>%
  ungroup() %>%
  ggplot(aes(x = norepinephrine, y = pred.mean)) +
  geom_jitter(aes(y = pred, col = norepinephrine),
              alpha = 0.07,
              data = data_test %>%
  mutate(pred = data_plot_gbm$y,
         pred.lower = data_plot_gbm$ymin,
         pred.upper = data_plot_gbm$ymax)) +
  geom_errorbar(aes(ymin = pred.low,
                      ymax = pred.up,
                    y = pred.mean), width = 0.1) +
  geom_point(size = 4) +
  theme_bw() +
  labs(title= "gbm",
       y = "Predicted probability of death after 72 hours from ICU admission",
       x = "Patient treated with norepinephrine",
       col = "") +
  scale_x_discrete(labels = c("No",
                              "Yes")) +
  theme(legend.position = "none") +
  scale_color_lancet()

#---

# 2. rf


# norepinephrine plots
data_test %>%
  mutate(pred = data_plot_rf$y,
         pred.lower = data_plot_rf$ymin,
         pred.upper = data_plot_rf$ymax) %>%
  dplyr::group_by(norepinephrine) %>%
  dplyr::summarise(pred.mean = mean(pred),
            pred.low = mean(pred.lower),
            pred.up = mean(pred.upper)) %>%
  ungroup() %>%
  ggplot(aes(x = norepinephrine, y = pred.mean)) +
  geom_jitter(aes(y = pred, col = norepinephrine),
              alpha = 0.07,
              data = data_test %>%
  mutate(pred = data_plot_rf$y,
         pred.lower = data_plot_rf$ymin,
         pred.upper = data_plot_rf$ymax)) +
  geom_errorbar(aes(ymin = pred.low,
                      ymax = pred.up,
                    y = pred.mean), width = 0.1) +
  geom_point(size = 4) +
  theme_bw() +
  labs(title= "rf",
       y = "Predicted probability of death after 72 hours from ICU admission",
       x = "Patient treated with norepinephrine",
       col = "") +
  scale_x_discrete(labels = c("No",
                              "Yes")) +
  theme(legend.position = "none") +
  scale_color_lancet()

#---

# 3. xgb


# norepinephrine plots
data_test %>%
  mutate(pred = data_plot_xgb$y,
         pred.lower = data_plot_xgb$ymin,
         pred.upper = data_plot_xgb$ymax) %>%
  dplyr::group_by(norepinephrine) %>%
  dplyr::summarise(pred.mean = mean(pred),
            pred.low = mean(pred.lower),
            pred.up = mean(pred.upper)) %>%
  ungroup() %>%
  ggplot(aes(x = norepinephrine, y = pred.mean)) +
  geom_jitter(aes(y = pred, col = norepinephrine),
              alpha = 0.07,
              data = data_test %>%
  mutate(pred = data_plot_xgb$y,
         pred.lower = data_plot_xgb$ymin,
         pred.upper = data_plot_xgb$ymax)) +
  geom_errorbar(aes(ymin = pred.low,
                      ymax = pred.up,
                    y = pred.mean), width = 0.1) +
  geom_point(size = 4) +
  theme_bw() +
  labs(title= "xgb",
       y = "Predicted probability of death after 72 hours from ICU admission",
       x = "Patient treated with norepinephrine",
       col = "") +
  scale_x_discrete(labels = c("No",
                              "Yes")) +
  theme(legend.position = "none") +
  scale_color_lancet()

```

## respiratoryrate_clear

```{r}

# 1. gbm


# respiratoryrate_clear
data_test %>%
  mutate(pred = data_plot_gbm$y,
         pred.lower = data_plot_gbm$ymin,
         pred.upper = data_plot_gbm$ymax) %>%
  dplyr::group_by(respiratoryrate_clear) %>%
  dplyr::summarise(pred.mean = mean(pred),
            pred.low = mean(pred.lower),
            pred.up = mean(pred.upper)) %>%
  ungroup() %>%
  ggplot(aes(x = respiratoryrate_clear, y = round(pred.mean, 2))) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred.mean = ifelse(icu_mortality_72hrs == "Expired", 1, 0))) +
  geom_point(alpha = 0.5) +
  geom_smooth(fill = "grey50",
              col = "black") +
  geom_vline(xintercept = 0, linetype = 2) +
  labs(title= "gbm",
       x = "Respiratory rate difference between 24 hours before ICU admission and ICU admission",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(-50, 100, 5)) +
  scale_y_continuous(breaks = seq(0, 1, 0.25),
                     limits = c(0, 1)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

#---

# 2. rf


# respiratoryrate_clear
data_test %>%
  mutate(pred = data_plot_rf$y,
         pred.lower = data_plot_rf$ymin,
         pred.upper = data_plot_rf$ymax) %>%
  dplyr::group_by(respiratoryrate_clear) %>%
  dplyr::summarise(pred.mean = mean(pred),
            pred.low = mean(pred.lower),
            pred.up = mean(pred.upper)) %>%
  ungroup() %>%
  ggplot(aes(x = respiratoryrate_clear, y = round(pred.mean, 2))) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred.mean = ifelse(icu_mortality_72hrs == "Expired", 1, 0))) +
  geom_point(alpha = 0.5) +
  geom_smooth(fill = "grey50",
              col = "black") +
  geom_vline(xintercept = 0, linetype = 2) +
  labs(title= "rf",
       x = "Respiratory rate difference between 24 hours before ICU admission and ICU admission",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(-50, 100, 5)) +
  scale_y_continuous(breaks = seq(0, 1, 0.25),
                     limits = c(0, 1)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

#---

# 3. xgb


# respiratoryrate_clear
data_test %>%
  mutate(pred = data_plot_xgb$y,
         pred.lower = data_plot_xgb$ymin,
         pred.upper = data_plot_xgb$ymax) %>%
  dplyr::group_by(respiratoryrate_clear) %>%
  dplyr::summarise(pred.mean = mean(pred),
            pred.low = mean(pred.lower),
            pred.up = mean(pred.upper)) %>%
  ungroup() %>%
  ggplot(aes(x = respiratoryrate_clear, y = round(pred.mean, 2))) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred.mean = ifelse(icu_mortality_72hrs == "Expired", 1, 0))) +
  geom_point(alpha = 0.5) +
  geom_smooth(fill = "grey50",
              col = "black") +
  geom_vline(xintercept = 0, linetype = 2) +
  labs(title= "xgb",
       x = "Respiratory rate difference between 24 hours before ICU admission and ICU admission",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(-50, 100, 5)) +
  scale_y_continuous(breaks = seq(0, 1, 0.25),
                     limits = c(0, 1)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

```

## temperature_clear

```{r}

# 1. gbm


# temperature_clear
data_test %>%
  mutate(pred = data_plot_gbm$y,
         pred.lower = data_plot_gbm$ymin,
         pred.upper = data_plot_gbm$ymax) %>%
  dplyr::group_by(temperature_clear) %>%
  dplyr::summarise(pred.mean = mean(pred),
            pred.low = mean(pred.lower),
            pred.up = mean(pred.upper)) %>%
  ungroup() %>%
  ggplot(aes(x = temperature_clear, y = round(pred.mean, 2))) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred.mean = ifelse(icu_mortality_72hrs == "Expired", 1, 0))) +
  geom_point(alpha = 0.5) +
  geom_smooth(fill = "grey50",
              col = "black") +
  geom_vline(xintercept = 0, linetype = 2) +
  labs(title= "gbm",
       x = "Temperature rate difference between 24 hours before ICU admission and ICU admission",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(-50, 100, 2.5)) +
  scale_y_continuous(breaks = seq(0, 1, 0.25),
                     limits = c(0, 1)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

#---

# 2. rf


# temperature_clear
data_test %>%
  mutate(pred = data_plot_rf$y,
         pred.lower = data_plot_rf$ymin,
         pred.upper = data_plot_rf$ymax) %>%
  dplyr::group_by(temperature_clear) %>%
  dplyr::summarise(pred.mean = mean(pred),
            pred.low = mean(pred.lower),
            pred.up = mean(pred.upper)) %>%
  ungroup() %>%
  ggplot(aes(x = temperature_clear, y = round(pred.mean, 2))) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred.mean = ifelse(icu_mortality_72hrs == "Expired", 1, 0))) +
  geom_point(alpha = 0.5) +
  geom_smooth(fill = "grey50",
              col = "black") +
  geom_vline(xintercept = 0, linetype = 2) +
  labs(title= "rf",
       x = "Temperature rate difference between 24 hours before ICU admission and ICU admission",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(-50, 100, 2.5)) +
  scale_y_continuous(breaks = seq(0, 1, 0.25),
                     limits = c(0, 1)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

#---

# 3. xgb


# temperature_clear
data_test %>%
  mutate(pred = data_plot_xgb$y,
         pred.lower = data_plot_xgb$ymin,
         pred.upper = data_plot_xgb$ymax) %>%
  dplyr::group_by(temperature_clear) %>%
  dplyr::summarise(pred.mean = mean(pred),
            pred.low = mean(pred.lower),
            pred.up = mean(pred.upper)) %>%
  ungroup() %>%
  ggplot(aes(x = temperature_clear, y = round(pred.mean, 2))) +
  geom_point(aes(col = icu_mortality_72hrs), alpha = 0.1,
             data = data_test %>%
               mutate(prev_rounded = round(prev_dept_los_hrs, 1),
                      pred.mean = ifelse(icu_mortality_72hrs == "Expired", 1, 0))) +
  geom_point(alpha = 0.5) +
  geom_smooth(fill = "grey50",
              col = "black") +
  geom_vline(xintercept = 0, linetype = 2) +
  labs(title= "xgb",
       x = "Temperature rate difference between 24 hours before ICU admission and ICU admission",
       y = "Predicted probability of death after 72 hours from ICU admission",
       col = "Mortality") +
  theme_minimal() +
  theme(legend.position = "bottom") +
  scale_x_continuous(breaks = seq(-50, 100, 2.5)) +
  scale_y_continuous(breaks = seq(0, 1, 0.25),
                     limits = c(0, 1)) +
  ggsci::scale_color_lancet() +
  guides(col = guide_legend(override.aes = list(alpha = 1)))

```

After observing the predictions and due to the smaller uncertainty interval (higher precision), also taking into account the F1 parameter, the Balanced Accuracy and the confusion matrices, if only one model had to be selected, the rf model would be selected.
